% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commit.r
\name{add_session_info}
\alias{add_session_info}
\title{Add sessionInfo to message}
\usage{
add_session_info(message)
}
\arguments{
\item{message}{The message.}
}
\value{
message with appended sessionInfo
}
\description{
Add sessionInfo to message
}
\keyword{internal}

