% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ahead_behind}
\alias{ahead_behind}
\alias{ahead_behind,git_commit,git_commit-method}
\title{Ahead Behind}
\usage{
ahead_behind(local, upstream)

\S4method{ahead_behind}{git_commit,git_commit}(local, upstream)
}
\arguments{
\item{local}{a S4 class git_commit \code{object}.}

\item{upstream}{a S4 class git_commit \code{object}.}
}
\value{
An integer vector of length 2 with number of commits that
the upstream commit is ahead and behind the local commit
}
\description{
Count the number of unique commits between two commit objects.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

ahead_behind(commits(repo)[[1]], commits(repo)[[2]])
}
}
\keyword{methods}

