% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{blame}
\alias{blame}
\alias{blame,git_repository,character-method}
\title{Get blame for file}
\usage{
blame(repo, path)

\S4method{blame}{git_repository,character}(repo, path)
}
\arguments{
\item{repo}{The repository}

\item{path}{Path to the file to consider}
}
\value{
S4 class git_blame object
}
\description{
Get blame for file
}
\keyword{methods}

