\docType{class}
\name{branch-class}
\alias{branch-class}
\alias{show,branch-method}
\title{Class \code{"branch"}}
\description{
  S4 class to handle a git branch, contains
  \code{\link{reference}}.
}
\section{Slots}{
  \describe{ \item{remote}{ :TODO:DOCUMENTATION: }
  \item{url}{ :TODO:DOCUMENTATION: } \item{head}{
  :TODO:DOCUMENTATION: } }
}

\section{Methods}{
  \describe{ \item{show}{\code{signature(object =
  "branch")}} }
}
\keyword{classes}
\keyword{methods}

