% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_create}
\alias{branch_create}
\alias{branch_create,git_commit-method}
\title{Create a branch}
\usage{
branch_create(commit, name, force = FALSE, message = NULL, who = NULL)

\S4method{branch_create}{git_commit}(commit, name, force = FALSE,
  message = NULL, who = NULL)
}
\arguments{
\item{commit}{Commit to which branch should point.}

\item{name}{Name for the branch}

\item{force}{Overwrite existing branch. Default = FALSE}

\item{message}{The one line long message to the reflog. Default is
NULL, which gives the log message "Branch: created"}

\item{who}{The identity that will be used to populate the
reflog entry. Default is NULL, which gives the default signature.}
}
\value{
invisible S4 class git_branch object
}
\description{
Create a branch
}
\keyword{methods}

