% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{branch_get_upstream}
\alias{branch_get_upstream}
\alias{branch_get_upstream,git_branch-method}
\title{Get remote tracking branch}
\usage{
branch_get_upstream(branch)

\S4method{branch_get_upstream}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
S4 class git_branch or NULL if no remote tracking branch.
}
\description{
Get remote tracking branch, given a local branch.
}
\keyword{methods}

