% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{branch_remote_name}
\alias{branch_remote_name}
\alias{branch_remote_name,git_branch-method}
\title{Remote name of a branch}
\usage{
branch_remote_name(branch)

\S4method{branch_remote_name}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
character string with remote name
}
\description{
The name of remote that the remote tracking branch belongs to
}
\examples{
\dontrun{
## Initialize two temporary repositories
path_bare <- tempfile(pattern="git2r-")
path_repo <- tempfile(pattern="git2r-")
dir.create(path_bare)
dir.create(path_repo)
repo_bare <- init(path_bare, bare = TRUE)
repo <- clone(path_bare, path_repo)

## Config user and commit a file
config(repo, user.name="Author", user.email="author@example.org")

## Write to a file and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           file.path(path_repo, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Push commits from repository to bare repository
## Adds an upstream tracking branch to branch 'master'
push(repo, "origin", "refs/heads/master")

## Get remote name
branch_remote_name(branches(repo)[[2]])
}
}
\keyword{methods}

