% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{branch_remote_url}
\alias{branch_remote_url}
\alias{branch_remote_url,git_branch-method}
\title{Remote url of a branch}
\usage{
branch_remote_url(branch)

\S4method{branch_remote_url}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
character string with remote url
}
\description{
Remote url of a branch
}
\keyword{methods}

