% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_set_upstream}
\alias{branch_set_upstream}
\alias{branch_set_upstream,git_branch-method}
\title{Set remote tracking branch}
\usage{
branch_set_upstream(branch, name)

\S4method{branch_set_upstream}{git_branch}(branch, name)
}
\arguments{
\item{branch}{The branch to configure}

\item{name}{remote-tracking or local branch to set as
upstream. Pass NULL to unset.}
}
\value{
invisible NULL
}
\description{
Set the upstream configuration for a given local branch
}
\keyword{methods}

