% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{branches-methods}
\alias{branches}
\alias{branches-methods}
\title{Branches}
\usage{
branches(object, flags = c("ALL", "LOCAL", "REMOTE"))
}
\arguments{
\item{object}{The repository}

\item{flags}{Filtering flags for the branch listing. Valid values
are 'ALL', 'LOCAL' or 'REMOTE'}
}
\value{
list of branches in repository
}
\description{
List branches in repository
}
\keyword{methods}

