% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{checkout}
\alias{checkout}
\alias{checkout,git_repository-method}
\title{Checkout}
\usage{
checkout(repo, treeish = "HEAD")

\S4method{checkout}{git_repository}(repo, treeish = "HEAD")
}
\arguments{
\item{repo}{The repository.}

\item{treeish}{a commit, tag or tree which content will be used to
update the working directory or the default to use 'HEAD'.}
}
\value{
invisible NULL
}
\description{
Update files in the and working tree to match the content of the
tree pointed at by the treeish.
}
\keyword{methods}

