% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clone}
\alias{clone}
\title{Clone a remote repository}
\usage{
clone(url, local_path, credentials = NULL, progress = TRUE)
}
\arguments{
\item{url}{The remote repository to clone}

\item{local_path}{Local directory to clone to}

\item{credentials}{The credentials for remote repository access. Default
is NULL.}

\item{progress}{Show progress. Default is TRUE.}
}
\value{
A S4 \code{git_repository} object
}
\description{
Clone a remote repository
}
\examples{
\dontrun{
## Clone a remote repository
repo <- clone("https://github.com/ropensci/git2r", "path/to/git2r")
}
}
\seealso{
\code{\linkS4class{cred_plaintext}},
\code{\linkS4class{cred_ssh_key}}
}
\keyword{methods}

