% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{coerce-git_repository-method}
\alias{coerce,git_repository,data.frame-method}
\alias{coerce-git_repository-method}
\title{Coerce Git repository to a \code{data.frame}}
\arguments{
\item{from}{The repository \code{object}}
}
\value{
\code{data.frame}
}
\description{
The commits in the repository are coerced to a \code{data.frame}
}
\details{
The \code{data.frame} have the following columns:
\describe{

\item{hex}{
    the SHA-1 hash as 40 characters of hexadecimal
  }

\item{summary}{
    the short "summary" of the git commit message.
  }

\item{message}{
    the full message of a commit
  }

\item{author}{
    full name of the author
  }

\item{email}{
    email of the author
  }

\item{when}{
    time when the commit happened
  }

}
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Coerce commits to a data.frame
df <- as(repo, "data.frame")

str(df)
}
}
\keyword{methods}

