% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{commits-methods}
\alias{commits}
\alias{commits-methods}
\title{Commits}
\usage{
commits(object)
}
\arguments{
\item{object}{The repository \code{object}.}
}
\value{
list of commits in repository
}
\description{
Commits
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository('path/to/git2r')

## List commits in repository
commits(repo)

## List commits direct from path
commits('path/to/git2r')
}
}
\keyword{methods}

