% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_find_global}
\alias{config_find_global}
\title{Locate the path to the global configuration file}
\usage{
config_find_global()
}
\value{
path if a global configuration file has been found, else
    NA.
}
\description{
The user or global configuration file is usually located in
'$HOME/.gitconfig'. This method will try to guess the full path to
that file, if the file exists. This method will not guess the path
to the xdg compatible config file (.config/git/config).
}
