% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{content}
\alias{content}
\alias{content,git_blob-method}
\title{Content of blob}
\usage{
content(blob, split = TRUE)

\S4method{content}{git_blob}(blob, split = TRUE)
}
\arguments{
\item{object}{The blob \code{object}.}

\item{split}{Split blob content to text lines. Default TRUE.}
}
\value{
The content of the blob
}
\description{
Content of blob
}
\keyword{methods}

