\name{contribution_wc}
\alias{contribution_wc}
\title{Contribution word cloud}
\usage{
contribution_wc(dir = getwd())
}
\arguments{
  \item{dir}{Path to git repo. Default is current working
  directory.}
}
\description{
Generate a word cloud from commit messages
}
\examples{
\dontrun{
contribution_wc()
}
}
\author{
Karthik Ram \email{karthik.ram@gmail.com}
}

