% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{contributions}
\alias{contributions}
\alias{contributions,character-method}
\alias{contributions,git_repository-method}
\alias{contributions,missing-method}
\title{Contributions}
\usage{
contributions(repo, breaks = c("month", "week", "day"), by = c("commits",
  "author"))

\S4method{contributions}{missing}(repo, breaks = c("month", "week", "day"),
  by = c("commits", "author"))

\S4method{contributions}{character}(repo, breaks = c("month", "week", "day"),
  by = c("commits", "author"))

\S4method{contributions}{git_repository}(repo, breaks = c("month", "week",
  "day"), by = c("commits", "author"))
}
\arguments{
\item{repo}{The repository.}

\item{breaks}{Default is \code{month}. Change to week or day as necessary.}

\item{by}{Contributions by "commits" or "author". Default is "commits".}
}
\value{
A \code{data.frame} with contributions.
}
\description{
See contributions to a Git repo
}
\examples{
\dontrun{
## If current working dir is a git repo
contributions()
contributions(breaks = "week")
contributions(breaks = "day")

## If the path is somewhere else
contributions("/path/to/repo")
}
}
\author{
Karthik Ram \email{karthik.ram@gmail.com}

Stefan Widgren \email{stefan.widgren@gmail.com}
}
\keyword{methods}

