% Generated by roxygen2 (4.0.0): do not edit by hand
\name{contributions}
\alias{contributions}
\title{Contributions_per_month}
\usage{
contributions(dir = getwd(), title = "Commits", breaks = "month",
  data_only = FALSE)
}
\arguments{
\item{dir}{Path to git repo. Default is current working directory.}

\item{title}{Default title for the plot is commits. Supply a new title if you desire one.}

\item{breaks}{Default is \code{month}. Change to week or day as necessary.}

\item{data_only}{= default is \code{FALSE}. Set to \code{TRUE} if you just require the \code{data.frame} output.}
}
\description{
Generates a ggplot chart of number of contributions to git repo per month
}
\examples{
\dontrun{
# If current working dir is a git repo
# contributions()
# contributions(breaks = "week")
# contributions(breaks = "day")
# If the path is somewhere else
# contributions(dir = "/path/to/repo")
# If you need just the data, not the plot
contributions(data_only = TRUE)
}
}
\author{
Karthik Ram \email{karthik.ram@gmail.com}
}

