% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{create_blob_fromdisk}
\alias{create_blob_fromdisk}
\alias{create_blob_fromdisk,git_repository,character-method}
\title{Create blob from file on disk}
\usage{
create_blob_fromdisk(repo, path)

\S4method{create_blob_fromdisk}{git_repository,character}(repo, path)
}
\arguments{
\item{repo}{The repository where the blob(s) will be written. Can
be a bare repository.}

\item{path}{The file(s) from which the blob will be created.}
}
\value{
list of S4 class git_blob \code{objects}
}
\description{
Read a file from the filesystem and write its content to the
Object Database as a loose blob. The method is vectorized and
accepts a vector of files to create blobs from.
}
\keyword{methods}

