% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{create_blob_fromworkdir}
\alias{create_blob_fromworkdir}
\alias{create_blob_fromworkdir,git_repository,character-method}
\title{Create blob from file in working directory}
\usage{
create_blob_fromworkdir(repo, relative_path)

\S4method{create_blob_fromworkdir}{git_repository,character}(repo,
  relative_path)
}
\arguments{
\item{repo}{The repository where the blob(s) will be
written. Cannot be a bare repository.}

\item{relative_path}{The file(s) from which the blob will be
created, relative to the repository's working dir.}
}
\value{
list of S4 class git_blob \code{objects}
}
\description{
Read a file from the working folder of a repository and write its
content to the Object Database as a loose blob. The method is
vectorized and accepts a vector of files to create blobs from.
}
\keyword{methods}

