% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{cred_env-class}
\alias{cred_env-class}
\title{S4 class to handle environmental variables credential
object}
\description{
Class \code{"cred_env"}
}
\section{Slots}{

\describe{
  \item{username}{
    The name of the environmental variable that holds
    the username for the authentication.
  }
  \item{password}{
    The name of the environmental variable that holds
    the password for the authentication.
  }
}
}

\keyword{classes}
\keyword{methods}
