% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{cred_plaintext}
\alias{cred_plaintext}
\alias{cred_plaintext,character,character-method}
\title{Create a new plain-text username and password credential object}
\usage{
cred_plaintext(username, password)

\S4method{cred_plaintext}{character,character}(username, password)
}
\arguments{
\item{username}{The username of the credential}

\item{password}{The password of the credential}
}
\value{
A S4 \code{cred_plaintext} object
}
\description{
Create a new plain-text username and password credential object
}
\keyword{methods}

