% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{cred_ssh_key-class}
\alias{cred_ssh_key-class}
\title{S4 class to handle a passphrase-protected ssh key
credential object}
\description{
Class \code{"cred_ssh_key"}
}
\section{Slots}{

\describe{
  \item{publickey}{
    The path to the public key of the credential
  }
  \item{privatekey}{
    The path to the private key of the credential
  }
  \item{passphrase}{
    The passphrase of the credential
  }
}
}
\keyword{classes}
\keyword{methods}

