% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{cred_token-class}
\alias{cred_token-class}
\title{S4 class to handle a personal access token credential
object}
\description{
Class \code{"cred_token"}
}
\section{Slots}{

\describe{
  \item{token}{
    The name of the environmental variable that holds
    the personal access token for the authentication.
  }
}
}
\keyword{classes}
\keyword{methods}

