\docType{methods}
\name{default_signature-methods}
\alias{default_signature}
\alias{default_signature,git_repository-method}
\alias{default_signature-methods}
\title{Get the signature}
\arguments{
  \item{object}{The repository \code{object} to check
  signature}
}
\value{
  Character vector with signature
}
\description{
  Get the signature according to the repository's
  configuration
}
\keyword{methods}

