% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{descendant_of}
\alias{descendant_of}
\alias{descendant_of,git_commit,git_commit-method}
\title{Descendant}
\usage{
descendant_of(commit, ancestor)

\S4method{descendant_of}{git_commit,git_commit}(commit, ancestor)
}
\arguments{
\item{commit}{a S4 class git_commit \code{object}.}

\item{ancestor}{a S4 class git_commit \code{object} to check if
ancestor to \code{commit}.}
}
\value{
TRUE if commit is descendant of \code{ancestor}, else
FALSE
}
\description{
Determine if a commit is the descendant of another commit
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

descendant_of(commits(repo)[[1]], commits(repo)[[2]])
}
}
\keyword{methods}

