% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{diff}
\alias{diff}
\alias{diff,git_repository-method}
\alias{diff,git_tree-method}
\title{Changes between commits, trees, working tree, etc.}
\usage{
diff(object, ...)

\S4method{diff}{git_repository}(object, index = FALSE)

\S4method{diff}{git_tree}(object, new_tree = NULL, index = FALSE)
}
\arguments{
\item{object}{A \code{git_repository} object or the old
\code{git_tree} object to compare to.}

\item{...}{Additional arguments affecting the diff produced}

\item{index}{\describe{
  \item{\emph{When object equals a git_repository}}{
    Whether to compare the index to HEAD. If FALSE (the default),
    then the working tree is compared to the index.
  }
  \item{\emph{When object equals a git_tree}}{
    Whether to use the working directory (by default), or the index
    (if set to TRUE) in the comparison to \code{object}.
  }
}}

\item{new_tree}{The new git_tree object to compare, or NULL.
If NULL, then we use the working directory or the index (see
the \code{index} argument).}
}
\value{
A \code{\linkS4class{git_diff}} object.
}
\description{
Changes between commits, trees, working tree, etc.
}
\examples{
\dontrun{
## Initialize a repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Config user
config(repo, user.name="User", user.email="user@example.org")

## Create a file, add, commit
writeLines("Hello world!", file.path(path, "test.txt"))
add(repo, "test.txt")
commit(repo, "Commit message")

## Change the file
writeLines(c("Hello again!", "Here is a second line", "And a third"),
           file.path(path, "test.txt"))

## diff between index and workdir
diff_1 <- diff(repo)
summary(diff_1)

## Diff between index and HEAD is empty
diff_2 <- diff(repo, index=TRUE)
summary(diff_2)

## Diff between tree and working dir, same as diff_1
diff_3 <- diff(tree(commits(repo)[[1]]))
summary(diff_3)

## Add changes, diff between index and HEAD is the same as diff_1
add(repo, "test.txt")
diff_4 <- diff(repo, index=TRUE)
summary(diff_4)

## Diff between tree and index
diff_5 <- diff(tree(commits(repo)[[1]]), index=TRUE)
summary(diff_5)

## Diff between two trees
commit(repo, "Second commit")
tree_1 <- tree(commits(repo)[[2]])
tree_2 <- tree(commits(repo)[[1]])
diff_6 <- diff(tree_1, tree_2)
summary(diff_6)

## Binary files
set.seed(42)
writeBin(as.raw((sample(0:255, 1000, replace=TRUE))),
         con=file.path(path, "test.bin"))
add(repo, "test.bin")
diff_7 <- diff(repo, index=TRUE)
summary(diff_7)
}
}
\keyword{methods}

