% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{diff_print}
\alias{diff_print}
\alias{diff_print,git_repository-method}
\alias{diff_print,git_tree-method}
\title{Print changes between commits, trees, working tree, etc.}
\usage{
diff_print(object, ...)

\S4method{diff_print}{git_repository}(object, index = FALSE,
  filename = NULL)

\S4method{diff_print}{git_tree}(object, new_tree = NULL, index = FALSE,
  filename = NULL)
}
\arguments{
\item{object}{A \code{git_repository} object or the old
\code{git_tree} object to compare to.}

\item{...}{Additional arguments affecting the diff_print produced}

\item{index}{\describe{
  \item{\emph{When object equals a git_repository}}{
    Whether to compare the index to HEAD. If FALSE (the default),
    then the working tree is compared to the index.
  }
  \item{\emph{When object equals a git_tree}}{
    Whether to use the working directory (by default), or the index
    (if set to TRUE) in the comparison to \code{object}.
  }
}}

\item{filename}{Determines where to write the diff. If filename
equals \code{NULL}, \code{character(0)}, \code{NA_character_} or
\code{""}, then the diff is written to a character vector. If
filename is a character vector of length one with non-NA value, the
diff is written to a file with name filename (the file is
overwritten if it exists). Default is NULL.}

\item{new_tree}{The new git_tree object to compare, or NULL.
If NULL, then we use the working directory or the index (see
the \code{index} argument).}
}
\value{
A character vector if filename equals \code{NULL},
\code{character(0)}, \code{NA_character_} or \code{""}. Oterwise
NULL.
}
\description{
Print changes between commits, trees, working tree, etc.
}
\keyword{methods}

