% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{discover_repository}
\alias{discover_repository}
\alias{discover_repository,character-method}
\title{Find path to repository for any file}
\usage{
discover_repository(path)

\S4method{discover_repository}{character}(path)
}
\arguments{
\item{path}{A character vector specifying the path to a file or folder}
}
\value{
Character vector with path to repository or NULL if this cannot be
  established.
}
\description{
libgit's git_discover_repository is used to identify the location of the
  repository. The path will therefore be terminated by a file separator.
}
\examples{
\dontrun{
path='/path/to/my/new/repo'
init(path)
discover_repository(path)
# /path/to/my/new/repo/.git/
}
}
\keyword{methods}

