% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_checkout}
\alias{do_checkout}
\title{Internal function to do the checkout}
\usage{
do_checkout(FUN, object, force, ref_log_head_spec)
}
\arguments{
\item{FUN}{Name of git2r C function to call}

\item{object}{The object to checkout}

\item{force}{If TRUE, then make working directory match
target. This will throw away local changes.}

\item{ref_log_target}{The target in the reflog message
in the reflog}
}
\description{
Internal function to do the checkout
}
\keyword{internal}

