% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{drop_stash}
\alias{drop_stash}
\alias{drop_stash,git_repository-method}
\alias{drop_stash,git_stash-method}
\title{Drop stash}
\usage{
drop_stash(object, ...)

\S4method{drop_stash}{git_repository}(object, index)

\S4method{drop_stash}{git_stash}(object)
}
\arguments{
\item{object}{The stash \code{object} to drop or a zero-based
integer to the stash to drop. The last stash has index 0.}
}
\value{
invisible NULL
}
\description{
Drop stash
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Assuming there are stashes in the repository.
## Drop a stash in repository.
drop_stash(stashes(repo)[[1]])

## Assuming there are stashes in the repository.
## Drop last stash in repository.
drop_stash(repo, 0)
}
}
\keyword{methods}

