% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fetch}
\alias{fetch}
\alias{fetch,git_repository-method}
\title{Fetch}
\usage{
fetch(repo, name)

\S4method{fetch}{git_repository}(repo, name)
}
\arguments{
\item{repo}{the repository}

\item{name}{the remote's name}
}
\value{
invisible(NULL)
}
\description{
Fetch
}
\keyword{methods}

