% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2r.R
\docType{package}
\name{git2r}
\alias{git2r}
\alias{git2r-package}
\title{git2r: R bindings to the libgit2 library}
\description{
git2r: R bindings to the libgit2 library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/git2r/ (website)}
  \item \url{https://github.com/ropensci/git2r}
  \item Report bugs at \url{https://github.com/ropensci/git2r/issues}
}

}
