% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_blame-class}
\alias{git_blame-class}
\title{S4 class to handle a git blame for a single file}
\description{
Class \code{"git_blame"}
}
\section{Slots}{

\describe{
  \item{path}{
    The path to the file of the blame
  }
  \item{hunks}{
    List of blame hunks
  }
  \item{repo}{
    The S4 class git_repository that contains the file
  }
}
}
\keyword{classes}
\keyword{methods}

