% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_blame_hunk-class}
\alias{git_blame_hunk-class}
\title{S4 class to handle a blame hunk}
\description{
Class \code{"git_blame_hunk"}
}
\section{Slots}{

\describe{
  \item{lines_in_hunk}{
    The number of lines in this hunk
  }
  \item{final_commit_id}{
    The hex of the commit where this line was last changed
  }
  \item{final_start_line_number}{
    The 1-based line number where this hunk begins, in the final
    version of the file
  }
  \item{final_signature}{
    :TODO:DOCUMENTATION:
  }
  \item{orig_commit_id}{
    :TODO:DOCUMENTATION:
  }
  \item{orig_start_line_number}{
    :TODO:DOCUMENTATION:
  }
  \item{orig_signature}{
    :TODO:DOCUMENTATION:
  }
  \item{orig_path}{
    :TODO:DOCUMENTATION:
  }
  \item{boundary}{
    :TODO:DOCUMENTATION:
  }
  \item{repo}{
    The S4 class git_repository that contains the blame hunk
  }
}
}
\keyword{classes}
\keyword{methods}

