\docType{class}
\name{git_branch-class}
\alias{git_branch-class}
\alias{show,git_branch-method}
\title{Class \code{"git_branch"}}
\description{
  S4 class to handle a git branch, contains
  \code{git_reference}.
}
\section{Slots}{
  \describe{ \item{remote}{ :TODO:DOCUMENTATION: }
  \item{url}{ :TODO:DOCUMENTATION: } \item{head}{
  :TODO:DOCUMENTATION: } }
}

\section{Methods}{
  \describe{ \item{is.head}{\code{signature(object =
  "git_branch")}} \item{is.local}{\code{signature(object =
  "git_branch")}} \item{show}{\code{signature(object =
  "git_branch")}} }
}
\keyword{classes}
\keyword{methods}

