\docType{class}
\name{git_commit-class}
\alias{git_commit-class}
\alias{show,git_commit-method}
\title{Class \code{"git_commit"}}
\description{
  S4 class to handle a git commit.
}
\section{Slots}{
  \describe{ \item{author}{ An author signature } }
}

\section{Methods}{
  \describe{ \item{show}{\code{signature(object =
  "git_commit")}} }
}
\keyword{classes}
\keyword{methods}

