% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_diff_file-class}
\alias{git_diff_file-class}
\title{Git diff file}
\description{
Git diff file
}
\section{Slots}{

\describe{
  \item{old_file}{
    :TODO:DOCUMENTATION:
  }
  \item{new_file}{
    :TODO:DOCUMENTATION:
  }
  \item{hunks}{
    List of \code{\linkS4class{git_diff_hunk}} objects.
  }
}
}
\keyword{classes}

