% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_diff_hunk-class}
\alias{git_diff_hunk-class}
\title{Git diff hunk}
\description{
Git diff hunk
}
\section{Slots}{

\describe{
  \item{old_start}{
    :TODO:DOCUMENTATION:
  }
  \item{old_lines}{
    :TODO:DOCUMENTATION:
  }
  \item{new_start}{
    :TODO:DOCUMENTATION:
  }
  \item{new_lines}{
    :TODO:DOCUMENTATION:
  }
  \item{header}{
    :TODO:DOCUMENTATION:
  }
  \item{lines}{
    :TODO:DOCUMENTATION:
  }
}
}
\keyword{classes}

