% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_diff_line-class}
\alias{git_diff_line-class}
\title{Git diff line}
\description{
Git diff line
}
\section{Slots}{

\describe{
  \item{origin}{
    :TODO:DOCUMENTATION:
  }
  \item{old_lineno}{
    :TODO:DOCUMENTATION:
  }
  \item{new_lineno}{
    :TODO:DOCUMENTATION:
  }
  \item{num_lines}{
    :TODO:DOCUMENTATION:
  }
  \item{content}{
    :TODO:DOCUMENTATION:
  }
}
}
\keyword{classes}

