% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{git_fetch_head-class}
\alias{git_fetch_head-class}
\title{S4 class to handle a fetch head}
\description{
Class \code{"git_fetch_head"}
}
\section{Slots}{

\describe{
  \item{ref_name}{
    The name of the ref.
  }
  \item{remote_url}{
    The url of the remote.
  }
  \item{sha}{
    The SHA-1 of the remote head that were updated during the last
    fetch.
  }
  \item{is_merge}{
    Is head for merge.
  }
  \item{repo}{
    The S4 class git_repository that contains the fetch head.
  }
}
}
\keyword{classes}
\keyword{methods}

