% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_merge_result-class}
\alias{git_merge_result-class}
\title{S4 class to handle the merge result}
\description{
Class \code{"git_merge_result"}
}
\section{Slots}{

\describe{
  \item{status}{
    The status of the merge.
  }
  \item{conflicts}{
    Does the index contain entries representing file conflicts.
  }
}
}
\keyword{classes}
\keyword{methods}

