% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_object-class}
\alias{git_object-class}
\title{S4 class to handle the objects in the database}
\description{
Class \code{"git_object"}
}
\section{Slots}{

\describe{
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1
  }
  \item{type}{
    The type of the object read from the header.
  }
  \item{len}{
    The length of the object read from the header.
  }
  \item{repo}{
    The S4 class git_repository that contains the blob
  }
}
}
\keyword{classes}
\keyword{methods}

