% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_recent.r
\docType{methods}
\name{git_recent}
\alias{git_recent}
\alias{git_recent,ANY,ANY-method}
\alias{git_recent,git_connection-methods}
\alias{git_recent,ANY,gitConnection-method}
\alias{git_recent,git_connection-methods}
\title{Get the info from the latest commit of a file}
\usage{
git_recent(file, connection, ...)

\S4method{git_recent}{ANY,ANY}(file, connection, ...)

\S4method{git_recent}{ANY,gitConnection}(file, connection, ...)
}
\arguments{
\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{gitConnection}
object}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path}
}
\description{
Get the info from the latest commit of a file
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
