% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_reference-class}
\alias{git_reference-class}
\title{S4 class to handle a git reference}
\description{
Class \code{"git_reference"}
}
\section{Slots}{

\describe{
  \item{name}{
    The full name of the reference.
  }
  \item{type}{
    Type of the reference, either direct (GIT_REF_OID == 1) or
    symbolic (GIT_REF_SYMBOLIC == 2)
  }
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1
  }
  \item{target}{
    The full name to the reference pointed to by a
    symbolic reference.
  }
  \item{shorthand}{
    The reference's short name
  }
}
}

\section{Methods}{

\describe{
  \item{show}{\code{signature(object = "git_reference")}}
}
}
\keyword{classes}
\keyword{methods}

