% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_reflog_entry-class}
\alias{git_reflog_entry-class}
\title{S4 class to handle a git reflog entry.}
\description{
Class \code{"git_reflog_entry"}
}
\section{Slots}{

\describe{
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1
  }
  \item{message}{
    The log message of the entry
  }
  \item{index}{
    The index (zero-based) of the entry in the reflog
  }
  \item{committer}{
    The committer signature
  }
  \item{refname}{
    Name of the reference
  }
  \item{repo}{
    The S4 class git_repository that contains the reflog entry
  }
}
}
\keyword{classes}
\keyword{methods}

