% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_sha.r
\docType{methods}
\name{git_sha}
\alias{git_sha}
\alias{git_sha,ANY,ANY-method}
\alias{git_sha,git_connection-methods}
\alias{git_sha,ANY,gitConnection-method}
\alias{git_sha,git_connection-methods}
\title{Get the SHA of the files at the HEAD}
\usage{
git_sha(file, connection, ...)

\S4method{git_sha}{ANY,ANY}(file, connection, ...)

\S4method{git_sha}{ANY,gitConnection}(file, connection, ...)
}
\arguments{
\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{gitConnection}
object}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path}
}
\description{
Get the SHA of the files at the HEAD
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
