\docType{class}
\name{git_signature-class}
\alias{git_signature-class}
\alias{show,git_signature-method}
\title{S4 class to handle a git signature}
\description{
S4 class to handle a git signature
}
\section{Slots}{
  \describe{ \item{name}{ The full name of the author.  }
  \item{email}{ Email of the author.  } \item{when}{ Time
  when the action happened.  } }
}
\keyword{classes}
\keyword{methods}

