% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{git_stash-class}
\alias{git_stash-class}
\title{S4 class to handle a git stash}
\description{
Class \code{"git_stash"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{message}{
    Optional description of the stash
  }
  \item{stasher}{
    The identity of the person performing the stash
  }
  \item{repo}{
    The S4 class git_repository that contains the stash
  }
}
}
\keyword{classes}
\keyword{methods}

