% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{git_tree-class}
\alias{git_tree-class}
\title{S4 class to handle a git tree}
\description{
Class \code{"git_tree"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
}
}
\keyword{classes}
\keyword{methods}

