% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{gitConnection-class}
\alias{gitConnection-class}
\title{The gitConnection class}
\description{
The gitConnection class
}
\section{Slots}{

  \describe{
   \item{\code{Repository}}{a git repository}
   \item{\code{LocalPath}}{a subdirectory wihtin the repository}
   \item{\code{Credentials}}{the credentials for the repository}
   \item{\code{CommitUser}}{the name of the user how will commit}
   \item{\code{CommitEmail}}{the email of the user how will commit}
  }
}

