% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{hash}
\alias{hash}
\title{Determine the sha from a blob string}
\usage{
hash(data = NULL)
}
\arguments{
\item{data}{The string vector to hash.}
}
\value{
A string vector with the sha for each string in data.
}
\description{
The blob is not written to the object database.
}
\examples{
\dontrun{
identical(hash(c("Hello, world!\n",
                 "test content\n")),
               c("af5626b4a114abcb82d63db7c8082c3c4756e51b",
                 "d670460b4b4aece5915caf5c68d12f560a9fe3e4"))
}
}
