% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hashfile}
\alias{hashfile}
\alias{hashfile,character-method}
\title{Determine the sha1 hex of a blob from a file}
\usage{
hashfile(path)

\S4method{hashfile}{character}(path)
}
\arguments{
\item{path}{The path vector with files to hash.}
}
\value{
A path vector with the sha1 hex for each file in path.
}
\description{
The blob is not written to the object database.
}
\examples{
\dontrun{
## Create a file. NOTE: The line endings from writeLines gives
## LF (line feed) on Unix/Linux and CRLF (carriage return, line feed)
## on Windows. The example use writeChar to have more control.
path <- tempfile()
f <- file(path, "wb")
writeChar("Hello, world!\\n", f, eos = NULL)
close(f)

## Generate hash
hashfile(path)
identical(hashfile(path), hash("Hello, world!\\n"))
}
}
\keyword{methods}

