\docType{methods}
\name{head-methods}
\alias{head}
\alias{head-methods}
\title{Get HEAD for a repo}
\usage{
\S4method{head}{git_repository}(x)
}
\arguments{
  \item{x}{The repository \code{x} to check head}
}
\value{
Character vector with head
}
\description{
Get HEAD for a repo
}
\keyword{methods}

