% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{init}
\alias{init}
\alias{init,character-method}
\title{Init a repository}
\usage{
init(path, bare = FALSE)

\S4method{init}{character}(path, bare = FALSE)
}
\arguments{
\item{path}{A path to where to init a git repository}

\item{bare}{If TRUE, a Git repository without a working directory
is created at the pointed path. If FALSE, provided path will be
considered as the working directory into which the .git directory
will be created.}
}
\value{
A S4 \code{\linkS4class{git_repository}} object
}
\description{
Init a repository
}
\examples{
\dontrun{
## Init a repository
repo <- init("path/to/repository-1")

## Init a bare repository
repo <- init("path/to/repository-2", bare = TRUE)
}
}
\seealso{
\link{repository}
}
\keyword{methods}

