% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{is.binary}
\alias{is.binary}
\alias{is.binary,git_blob-method}
\title{Is blob binary}
\usage{
is.binary(object)

\S4method{is.binary}{git_blob}(object)
}
\arguments{
\item{object}{The blob \code{object}.}
}
\value{
TRUE if binary data, FALSE if not.
}
\description{
Is blob binary
}
\keyword{methods}

