% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.blob}
\alias{is.blob}
\title{Check if object is S4 class git_blob}
\usage{
is.blob(object)
}
\arguments{
\item{object}{Check if object is S4 class git_blob}
}
\value{
TRUE if object is S4 class git_blob, else FALSE
}
\description{
Check if object is S4 class git_blob
}
\keyword{methods}

