% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{is.detached}
\alias{is.detached}
\alias{is.detached,git_repository-method}
\title{Check if HEAD of repository is detached}
\usage{
is.detached(object)

\S4method{is.detached}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object}}
}
\value{
TRUE if repository HEAD is detached, else FALSE
}
\description{
Check if HEAD of repository is detached
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Check if repository HEAD is detached
is.detached(repo)
}
}
\keyword{methods}

