% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_git.R
\name{is.git}
\alias{is.git}
\title{Is the path a git repository
Checks if a '.git' subdirectory exists in 'path'}
\usage{
is.git(path)
}
\arguments{
\item{path}{the path to check}
}
\value{
A logical vector with the same length as path
}
\description{
Is the path a git repository
Checks if a '.git' subdirectory exists in 'path'
}
