\docType{methods}
\name{is.head-methods}
\alias{is.head}
\alias{is.head,git_branch-method}
\alias{is.head-methods}
\title{Check if branch is head}
\arguments{
  \item{object}{The branch \code{object} to check if it's
  head}
}
\value{
  TRUE if branch is head, else FALSE
}
\description{
  Check if branch is head
}
\keyword{methods}

