% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_bare}
\alias{is_bare}
\alias{is_bare,git_repository-method}
\title{Check if repository is bare}
\usage{
is_bare(repo)

\S4method{is_bare}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository to check if it's bare}
}
\value{
TRUE if bare repository, else FALSE
}
\description{
Check if repository is bare
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Check if it's a bare repository
is_bare(repo)
}
}
\keyword{methods}

