% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_binary}
\alias{is_binary}
\alias{is_binary,git_blob-method}
\title{Is blob binary}
\usage{
is_binary(blob)

\S4method{is_binary}{git_blob}(blob)
}
\arguments{
\item{blob}{The blob \code{object}.}
}
\value{
TRUE if binary data, FALSE if not.
}
\description{
Is blob binary
}
\keyword{methods}

