% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_commit}
\alias{is_commit}
\title{Check if object is S4 class git_commit}
\usage{
is_commit(object)
}
\arguments{
\item{object}{Check if object is S4 class git_commit}
}
\value{
TRUE if object is S4 class git_commit, else FALSE
}
\description{
Check if object is S4 class git_commit
}
\keyword{methods}

