% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repository.R
\name{is_data_repo}
\alias{is_data_repo}
\title{Check if object is a data repository}
\usage{
is_data_repo(object)
}
\arguments{
\item{object}{the object to check}
}
\value{
TRUE is a data \code{git_repository}, else FALSE
}
\description{
Check if object is a data repository
}
\seealso{
repo init
}
