% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{is_detached}
\alias{is_detached}
\alias{is_detached,git_repository-method}
\title{Check if HEAD of repository is detached}
\usage{
is_detached(repo)

\S4method{is_detached}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository \code{object}}
}
\value{
TRUE if repository HEAD is detached, else FALSE
}
\description{
Check if HEAD of repository is detached
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Check if repository HEAD is detached
is_detached(repo)
}
}
\keyword{methods}

