% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_local}
\alias{is_local}
\alias{is_local,git_branch-method}
\title{Check if branch is local}
\usage{
is_local(object)

\S4method{is_local}{git_branch}(object)
}
\arguments{
\item{object}{The branch \code{object} to check if it's local}
}
\value{
TRUE if branch is local, else FALSE
}
\description{
Check if branch is local
}
\keyword{methods}

