% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_merge}
\alias{is_merge}
\alias{is_merge,git_commit-method}
\title{Is merge}
\usage{
is_merge(commit)

\S4method{is_merge}{git_commit}(commit)
}
\arguments{
\item{commit}{a S4 class git_commit \code{object}.}
}
\value{
TRUE if commit has more than one parent, else FALSE
}
\description{
Determine if a commit is a merge commit, i.e. has more than one
parent.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Display list of merge commits in repository
invisible(lapply(commits(repo)[sapply(commits(repo), is_merge)], show))
}
}
\keyword{methods}

