% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_shallow}
\alias{is_shallow}
\alias{is_shallow,git_repository-method}
\title{Determine if the repository was a shallow clone}
\usage{
is_shallow(repo)

\S4method{is_shallow}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository}
}
\value{
TRUE if shallow clone, else FALSE
}
\description{
Determine if the repository was a shallow clone
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Check if it's a shallow clone
is_shallow(repo)
}
}
\keyword{methods}

