% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{last_commit}
\alias{last_commit}
\title{Last commit}
\usage{
last_commit(repo = NULL)
}
\arguments{
\item{repo}{repository \code{object}
\code{\linkS4class{git_repository}}, or a path to a
repository, or \code{NULL}.  If the \code{repo} argument is
\code{NULL}, the repository is searched for with
\code{\link{discover_repository}} in the current working
directory.}
}
\description{
Get last commit in the current branch.
}
\examples{
\dontrun{
## Initialize a repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Config user
config(repo, user.name="Alice", user.email="alice@example.org")

## Write to a file and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Get last commit
last_commit(repo)
last_commit(path)
}
}
