% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff.r
\docType{methods}
\name{length,git_diff-method}
\alias{length,git_diff-method}
\title{Number of files in git_diff object}
\usage{
\S4method{length}{git_diff}(x)
}
\arguments{
\item{x}{The git_diff \code{object}}
}
\value{
a non-negative integer
}
\description{
Number of files in git_diff object
}
\keyword{#'}
\keyword{methods}

