% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,git_tree-method}
\alias{length,git_tree-method}
\title{Number of entries in tree}
\usage{
\S4method{length}{git_tree}(x)
}
\arguments{
\item{x}{The tree \code{object}}
}
\value{
a non-negative integer or double (which will be rounded
down)
}
\description{
Number of entries in tree
}
\keyword{methods}

