% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{length.git_diff}
\alias{length.git_diff}
\title{Number of files in git_diff object}
\usage{
\method{length}{git_diff}(x)
}
\arguments{
\item{x}{The git_diff \code{object}}
}
\value{
a non-negative integer
}
\description{
Number of files in git_diff object
}
