% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{length.git_tree}
\alias{length.git_tree}
\title{Number of entries in tree}
\usage{
\method{length}{git_tree}(x)
}
\arguments{
\item{x}{The tree \code{object}}
}
\value{
a non-negative integer or double (which will be rounded
down)
}
\description{
Number of entries in tree
}
