% Generated by roxygen2 (4.0.1): do not edit by hand
\name{libgit2_features}
\alias{libgit2_features}
\title{Compile time options for libgit2.}
\usage{
libgit2_features()
}
\value{
A list with threads, https and ssh set to TRUE/FALSE.
}
\description{
Compile time options for libgit2.
}
\examples{
\dontrun{
libgit2_features()
}
}
\keyword{methods}

