% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgit2.R
\name{libgit2_sha}
\alias{libgit2_sha}
\title{SHA of the libgit2 library}
\usage{
libgit2_sha()
}
\value{
The 40 character hexadecimal string of the SHA-1
}
\description{
SHA of the libgit2 library that the bundled source code is based
on
}
\examples{
libgit2_sha()
}
\keyword{methods}
