% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_git.r
\docType{methods}
\name{list_files_git}
\alias{list_files_git}
\alias{list_files_git,ANY-method}
\alias{list_files_git,git_connection-methods}
\alias{list_files_git,git_connection-method}
\alias{list_files_git,git_connection-methods}
\title{List the files in a path of a git repository}
\usage{
list_files_git(connection, pattern = NULL, full.names = FALSE, ...)

\S4method{list_files_git}{ANY}(connection, pattern = NULL,
  full.names = FALSE, ...)

\S4method{list_files_git}{git_connection}(connection, pattern = NULL,
  full.names = FALSE, ...)
}
\arguments{
\item{connection}{The path of a git repository or a \code{git_connection}
object}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path}
}
\description{
List the files in a path of a git repository
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
