% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{lookup}
\alias{lookup}
\alias{lookup,git_repository,character-method}
\title{Lookup}
\usage{
lookup(repo, sha)

\S4method{lookup}{git_repository,character}(repo, sha)
}
\arguments{
\item{repo}{The repository.}

\item{sha}{The identity of the object to lookup. Must be 4 to 40
characters long.}
}
\value{
a \code{git_blob} or \code{git_commit} or \code{git_tag}
or \code{git_tree} object
}
\description{
Lookup one object in a repository.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Lookup commit in repository
lookup(repo, "6fd0b22")
}
}
\keyword{methods}

