% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.r
\name{lookup_repository}
\alias{lookup_repository}
\title{Lookup repository for methods with missing repo argument}
\usage{
lookup_repository()
}
\value{
S4 class git_repository
}
\description{
Lookup repository for methods with missing repo argument
}
\keyword{internal}
