% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remote.r
\docType{methods}
\name{ls_remote}
\alias{ls_remote}
\alias{ls_remote,character,git_repository-method}
\alias{ls_remote,character,missing-method}
\title{List references in a remote repository}
\usage{
ls_remote(name, repo)

\S4method{ls_remote}{character,missing}(name)

\S4method{ls_remote}{character,git_repository}(name, repo)
}
\arguments{
\item{name}{Character vector with the with "remote" repository URL to query.}

\item{repo}{an optional repository object used to store data temporarily.}
}
\value{
Character vector for each reference with the associated commit IDs.
}
\description{
Displays references available in a remote repository along with the associated commit IDs.
}
\examples{
\dontrun{
ls_remote("https://github.com/ropensci/git2r")
}
}
\keyword{methods}

