% Generated by roxygen2 (4.0.2): do not edit by hand
\name{markdown_link}
\alias{markdown_link}
\title{Generate a markdown link for current commit}
\usage{
markdown_link(repo = ".")
}
\arguments{
\item{repo}{Path to local git repository}
}
\description{
This function would be ideal to use when writing manscripts with markdown.
By adding a bit of text like "All the files at the current state of the manuscript can by found in commit" followed by \code{markdown_link(repo)},
One can maintain some degree of provenance.
}
\examples{
\dontrun{
markdown_link(".") # or path to repo
}
}

