% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/merge.r
\name{merge_branch}
\alias{merge_branch}
\title{Merge branch into HEAD}
\usage{
merge_branch(branch, commit_on_success, merger)
}
\arguments{
\item{branch}{The branch}

\item{commit_on_success}{If there are no conflicts written to the
index, the merge commit will be committed.}

\item{merger}{Who made the merge.}
}
\value{
A \code{\linkS4class{git_merge_result}} object.
}
\description{
Merge branch into HEAD
}
\keyword{internal}

