% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge.r
\name{merge_named_branch}
\alias{merge_named_branch}
\title{Merge named branch into HEAD}
\usage{
merge_named_branch(repo, branch, commit_on_success, merger)
}
\arguments{
\item{repo}{The repository}

\item{branch}{Name of branch}

\item{commit_on_success}{If there are no conflicts written to the
index, the merge commit will be committed.}

\item{merger}{Who made the merge.}
}
\value{
A \code{\linkS4class{git_merge_result}} object.
}
\description{
Merge named branch into HEAD
}
\keyword{internal}

