% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repository.R
\name{meta}
\alias{meta}
\title{Optimise a vector for storage in to a git repository and add meta data}
\usage{
meta(x)
}
\arguments{
\item{x}{the vector}
}
\description{
Optimise a vector for storage in to a git repository and add meta data
}
\details{
\itemize{
   \item \code{meta.character} checks for the presence of \code{'NA'}.
   Because \code{\link{write_delim_git}} stores the data unquoted,
   \code{'NA'} and \code{NA} result in the same value in the file. Hence
   \code{\link{read_delim_git}} would report \code{'NA'} as \code{NA}.
   Therefore \code{meta.character} will throw an error with \code{'NA'} is
   detected.
}
}
