% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{note_create}
\alias{note_create}
\alias{note_create,git_commit-method}
\title{Add note for a object}
\usage{
note_create(object, message, ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)

\S4method{note_create}{git_commit}(object, message,
  ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)
}
\arguments{
\item{object}{The object to annotate}

\item{message}{Content of the note to add}

\item{ref}{Canonical name of the reference to use. Default is
\code{note_default_ref}.}

\item{author}{Signature of the notes note author}

\item{committer}{Signature of the notes note committer}

\item{force}{Overwrite existing note. Default is FALSE}
}
\value{
S4 class git_note
}
\description{
Add note for a object
}
\keyword{methods}

