% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{note_default_ref}
\alias{note_default_ref}
\alias{note_default_ref,git_repository-method}
\title{Default notes reference}
\usage{
note_default_ref(repo)

\S4method{note_default_ref}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository}
}
\value{
Character vector of length one with name of default notes
reference
}
\description{
Get the default notes reference for a repository
}
\keyword{methods}

