% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{odb_list}
\alias{odb_list}
\alias{odb_list,git_repository-method}
\title{List all objects available in the database}
\usage{
odb_list(repo)

\S4method{odb_list}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository}
}
\value{
list of S4 class git_object \code{objects}
}
\description{
List all objects available in the database
}
\keyword{methods}

