% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{parents}
\alias{parents}
\alias{parents,git_commit-method}
\title{Parents}
\usage{
parents(object)

\S4method{parents}{git_commit}(object)
}
\arguments{
\item{object}{a S4 class git_commit \code{object}.}
}
\value{
list of S4 git_commit objects
}
\description{
Get parents of a commit.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

parents(commits(repo)[[1]])
}
}
\keyword{methods}

