% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{plot,git_repository-method}
\alias{plot}
\alias{plot,git_repository-method}
\title{Plot contributions over time}
\usage{
\S4method{plot}{git_repository}(x, breaks = c("month", "week", "day"),
  by = c("commits", "user"), title = "Commits")
}
\arguments{
\item{x}{The repository to plot}

\item{breaks}{Default is \code{month}. Change to week or day as
necessary.}

\item{by}{Contributions by 'commits' or 'user'. Default is
'commits'.}
}
\value{
A \code{ggplot} chart
}
\description{
Plot contributions over time
}
\keyword{methods}

