% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{pull}
\alias{pull}
\alias{pull,git_repository-method}
\title{Pull}
\usage{
pull(repo)

\S4method{pull}{git_repository}(repo)
}
\arguments{
\item{repo}{the repository}
}
\value{
invisible(NULL)
}
\description{
Pull
}
\keyword{methods}

