% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{punch_card}
\alias{punch_card}
\alias{punch_card,git_repository-method}
\title{Punch card}
\usage{
punch_card(object)

\S4method{punch_card}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object}.}
}
\value{
A \code{ggplot} chart
}
\description{
Punch card
}
\keyword{methods}

