% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{push}
\alias{push}
\alias{push,git_repository-method}
\title{Push}
\usage{
push(repo, refspec)

\S4method{push}{git_repository}(repo, refspec)
}
\arguments{
\item{repo}{the repository}

\item{refspec}{the refspec to be pushed}
}
\value{
invisible(NULL)
}
\description{
Push
}
\keyword{methods}

