% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_git.R
\name{read_delim_file}
\alias{read_delim_file}
\title{Read a \code{data.frame} from a git repository}
\usage{
read_delim_file(file, repo)
}
\arguments{
\item{file}{the name of the file with file extension. Can include a path
relative to the path of the \code{repo}}

\item{repo}{a \code{git_repository} object, created with
\code{\link{repository}}}
}
\description{
Read a \code{data.frame} from a git repository
}
