% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_git.r
\docType{methods}
\name{read_delim_git}
\alias{read_delim_git}
\alias{read_delim_git,ANY,ANY-method}
\alias{read_delim_git,git_connection-methods}
\alias{read_delim_git,ANY,gitConnection-method}
\alias{read_delim_git,git_connection-methods}
\title{Read a tab delimited file from a git repository}
\usage{
read_delim_git(file, connection, ...)

\S4method{read_delim_git}{ANY,ANY}(file, connection, ...)

\S4method{read_delim_git}{ANY,gitConnection}(file, connection, ...)
}
\arguments{
\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{gitConnection}
object}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path}
}
\description{
Read a tab delimited file from a git repository
}
