% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repository.R
\name{recent_commit}
\alias{recent_commit}
\title{Recent file change
Retrieve the most recent commit in which a file or data object was changed.}
\usage{
recent_commit(repo, path = NULL, data = FALSE)
}
\arguments{
\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{path}{the path to the file or the data object. File extensions are silently ignored in case of a data object}

\item{data}{refers path to a file (FALSE) or a data object (TRUE). Defaults to FALSE}
}
\value{
A data.frame with commit, author and timestamp. Will contain multiple rows when multiple commits are made within the same second
}
\description{
Recent file change
Retrieve the most recent commit in which a file or data object was changed.
}
