% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R, R/repository.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{diff}
\alias{reexports}
\alias{head}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{base}{\code{\link[base]{diff}}}

  \item{utils}{\code{\link[utils]{head}}}
}}

