\docType{methods}
\name{references-methods}
\alias{references}
\alias{references-methods}
\alias{references,repository-method}
\title{Get all references that can be found in a repository.}
\arguments{
  \item{object}{The repository \code{object}.}
}
\value{
  Character vector with references
}
\description{
  Get all references that can be found in a repository.
}
\keyword{methods}

