% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reflog}
\alias{reflog}
\alias{reflog,git_repository,character-method}
\alias{reflog,git_repository,missing-method}
\title{List and view reflog information}
\usage{
reflog(repo, refname)

\S4method{reflog}{git_repository,missing}(repo)

\S4method{reflog}{git_repository,character}(repo, refname)
}
\arguments{
\item{repo}{S4 class git_repository \code{object}.}

\item{refname}{The name of the reference to list. HEAD by
default.}
}
\value{
invisible list of S4 git_reflog_entry objects
}
\description{
List and view reflog information
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

reflog(repo)
}
}
\keyword{methods}

