% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remote.r
\docType{methods}
\name{remote_ls}
\alias{remote_ls}
\alias{remote_ls,character-method}
\title{List references in a remote repository}
\usage{
remote_ls(name, repo = git2r::init(tempdir()), credentials = NULL)

\S4method{remote_ls}{character}(name, repo = git2r::init(tempdir()),
  credentials = NULL)
}
\arguments{
\item{name}{Character vector with the "remote" repository URL to query or
the name of the remote if a \code{repo} argument is given.}

\item{repo}{an optional repository object used if remotes are specified by name.}

\item{credentials}{The credentials for the remote repository.}
}
\value{
Character vector for each reference with the associated commit IDs.
}
\description{
Displays references available in a remote repository along with the
associated commit IDs.  Akin to the 'git ls-remote' command.
}
\examples{
\dontrun{
remote_ls("https://github.com/ropensci/git2r")
}
}
\keyword{methods}

