% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remote_remove}
\alias{remote_remove}
\alias{remote_remove,git_repository-method}
\title{Remove a remote}
\usage{
remote_remove(object, ...)

\S4method{remote_remove}{git_repository}(object, name)
}
\arguments{
\item{object}{The repository to work on}

\item{name}{The name of the remote to remove}
}
\value{
NULL, invisibly
}
\description{
Remove a remote
}
\keyword{methods}

