% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remote_rename}
\alias{remote_rename}
\alias{remote_rename,git_repository-method}
\title{Rename a remote}
\usage{
remote_rename(object, ...)

\S4method{remote_rename}{git_repository}(object, oldname, newname)
}
\arguments{
\item{object}{The repository in which the remote should be renamed.}

\item{oldname}{Old name of the remote}

\item{newname}{New name of the remote}
}
\value{
NULL, invisibly
}
\description{
Rename a remote
}
\keyword{methods}

