% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{remotes}
\alias{remotes}
\alias{remotes,git_repository-method}
\title{Get the configured remotes for a repo}
\usage{
remotes(repo)

\S4method{remotes}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository to get remotes from}
}
\value{
Character vector with remotes
}
\description{
Get the configured remotes for a repo
}
\keyword{methods}

