% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.r
\name{remotes}
\alias{remotes}
\title{Get the configured remotes for a repo}
\usage{
remotes(repo = NULL)
}
\arguments{
\item{repo}{a \code{\linkS4class{git_repository}} object, or a
path to a repository, or \code{NULL}.  If the \code{repo}
argument is \code{NULL}, the repository is searched for with
\code{\link{discover_repository}} in the current working
directory.}
}
\value{
Character vector with remotes
}
\description{
Get the configured remotes for a repo
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user and commit a file
config(repo, user.name="Alice", user.email="alice@example.org")
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Add a remote
remote_add(repo, "playground", "https://example.org/git2r/playground")
remotes(repo)
remote_url(repo, "playground")

## Rename a remote
remote_rename(repo, "playground", "foobar")
remotes(repo)
remote_url(repo, "foobar")

## Remove a remote
remote_remove(repo, "foobar")
remotes(repo)
}
}
