% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_files_git.r
\docType{methods}
\name{remove_files_git}
\alias{remove_files_git}
\alias{remove_files_git,ANY-method}
\alias{remove_files_git,git_connection-methods}
\alias{remove_files_git,git_connection-method}
\alias{remove_files_git,git_connection-methods}
\title{Remove all the files in a path of a git repository}
\usage{
remove_files_git(connection, pattern = NULL, ...)

\S4method{remove_files_git}{ANY}(connection, pattern = NULL, ...)

\S4method{remove_files_git}{git_connection}(connection, pattern = NULL, ...)
}
\arguments{
\item{connection}{The path of a git repository or a \code{git_connection}
object}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path}
}
\description{
Remove all the files in a path of a git repository
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
