\docType{class}
\name{repository-class}
\alias{repository-class}
\alias{show,repository-method}
\alias{summary,repository-method}
\title{Class \code{"repository"}}
\description{
  S4 class to handle a git repository
}
\section{Slots}{
  \describe{ \item{repo}{ External pointer to a git
  repository } }
}

\section{Methods}{
  \describe{ \item{is.bare}{\code{signature(object =
  "repository")}} \item{is.empty}{\code{signature(object =
  "repository")}} \item{show}{\code{signature(object =
  "repository")}} \item{summary}{\code{signature(object =
  "repository")}} }
}
\keyword{classes}
\keyword{methods}

