% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{repository_discover}
\alias{repository_discover}
\alias{repository_discover,character-method}
\title{Find path to repository for any file}
\usage{
repository_discover(object)

\S4method{repository_discover}{character}(object)
}
\arguments{
\item{object}{A character vector specifying the path to a file or folder}
}
\value{
Character vector with path to repository or NULL if this cannot be
  established.
}
\description{
libgit's git_repository_discover is used to identify the location of the
  repository. The path will therefore be terminated by a file separator.
}
\examples{
\dontrun{
path='/path/to/my/new/repo'
init(path)
repository_discover(path)
# /path/to/my/new/repo/.git/
}
}
\keyword{methods}

