% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repository.R
\name{rm_data}
\alias{rm_data}
\title{Remove data files
Remove all tsv and/or yml files within the path}
\usage{
rm_data(repo = ".", path = NULL, type = c("tsv", "yml", "both"),
  recursive = TRUE, stage = FALSE)
}
\arguments{
\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{path}{the directory in which to clean all the data files}

\item{type}{which file type should be removed}

\item{recursive}{remove files in subdirectories too}

\item{stage}{immediatly stage the changes}
}
\description{
Remove data files
Remove all tsv and/or yml files within the path
}
