% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{methods}
\name{sha}
\alias{sha}
\alias{sha,git_blob-method}
\alias{sha,git_commit-method}
\alias{sha,git_note-method}
\alias{sha,git_reference-method}
\alias{sha,git_reflog_entry-method}
\alias{sha,git_tag-method}
\alias{sha,git_tree-method}
\alias{sha,git_fetch_head-method}
\alias{sha,git_merge_result-method}
\title{Get the SHA-1 of a git object}
\usage{
sha(object)

\S4method{sha}{git_blob}(object)

\S4method{sha}{git_commit}(object)

\S4method{sha}{git_note}(object)

\S4method{sha}{git_reference}(object)

\S4method{sha}{git_reflog_entry}(object)

\S4method{sha}{git_tag}(object)

\S4method{sha}{git_tree}(object)

\S4method{sha}{git_fetch_head}(object)

\S4method{sha}{git_merge_result}(object)
}
\arguments{
\item{object}{a git object to get the SHA-1 from.}
}
\value{
The 40 character hexadecimal string of the SHA-1.
}
\description{
Get the 40 character hexadecimal string of the SHA-1.
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)
config(repo, user.name="Alice", user.email="alice@example.org")

## Create a file, add and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           con = file.path(path, "test.txt"))
add(repo, "test.txt")
sha(commit(repo, "Commit message 1"))
}
}
\keyword{methods}
