% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_blob-method}
\alias{show,git_blob-method}
\alias{show,git_blob-methods}
\title{Brief summary of blob}
\usage{
\S4method{show}{git_blob}(object)
}
\arguments{
\item{object}{The blob \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of blob
}
\keyword{methods}

