% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{show,git_branch-method}
\alias{show,git_branch-method}
\alias{show,git_branch-methods}
\title{Brief summary of branch}
\usage{
\S4method{show}{git_branch}(object)
}
\arguments{
\item{object}{The branch \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of branch
}
\keyword{methods}

