% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_merge_result-method}
\alias{show,git_merge_result-method}
\alias{show,git_merge_result-methods}
\title{Brief summary of merge result}
\usage{
\S4method{show}{git_merge_result}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{git_merge_result}} \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of merge result
}
\keyword{methods}

