% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/note.r
\docType{methods}
\name{show,git_note-method}
\alias{show,git_note-method}
\alias{show,git_note-methods}
\title{Brief summary of note}
\usage{
\S4method{show}{git_note}(object)
}
\arguments{
\item{object}{The \linkS4class{git_note} \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of note
}
\examples{
\dontrun{
## Create and initialize a repository in a temporary directory
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)
config(repo, user.name="Alice", user.email="alice@example.org")

## Create a file, add and commit
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Create note
note_create(commits(repo)[[1]], "My note")
}
}
\keyword{methods}

