% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_reference-method}
\alias{show,git_reference-method}
\alias{show,git_reference-methods}
\title{Brief summary of reference}
\usage{
\S4method{show}{git_reference}(object)
}
\arguments{
\item{object}{The reference \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of reference
}
\keyword{methods}

