% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,git_reflog_entry-method}
\alias{show,git_reflog_entry-method}
\alias{show,git_reflog_entry-methods}
\title{Brief summary of a reflog entry}
\usage{
\S4method{show}{git_reflog_entry}(object)
}
\arguments{
\item{object}{The reflog entry \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of a reflog entry
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## View repository HEAD reflog
reflog(repo)
}
}
\keyword{methods}

