\docType{methods}
\name{show,git_signature-method}
\alias{show,git_signature-method}
\alias{show,git_signature-methods}
\title{Brief summary of signature}
\usage{
\S4method{show}{git_signature}(object)
}
\arguments{
  \item{object}{The repository \code{object}}
}
\value{
:TODO:DOCUMENTATION:
}
\description{
Brief summary of signature
}
\keyword{methods}

