% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,git_stash-method}
\alias{show,git_stash-method}
\alias{show,git_stash-methods}
\title{Brief summary of a stash}
\usage{
\S4method{show}{git_stash}(object)
}
\arguments{
\item{object}{The stash \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of a stash
}
\keyword{methods}

