% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_tag-method}
\alias{show,git_tag-method}
\alias{show,git_tag-methods}
\title{Brief summary of a tag}
\usage{
\S4method{show}{git_tag}(object)
}
\arguments{
\item{object}{The tag \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of a tag
}
\keyword{methods}

