% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{show,git_tree-method}
\alias{show,git_tree-method}
\alias{show,git_tree-methods}
\title{Brief summary of tree}
\usage{
\S4method{show}{git_tree}(object)
}
\arguments{
\item{object}{The tree \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of tree
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Brief summary of one tree in repository
tree(commits(repo)[[1]])
}
}
\keyword{methods}

