\docType{methods}
\name{show}
\alias{show}
\alias{show,git_repository-methods}
\alias{show,git_signature-methods}
\title{Brief summary of repository}
\arguments{
  \item{object}{The repository \code{object}}

  \item{object}{The repository \code{object}}
}
\value{
  :TODO:DOCUMENTATION:

  :TODO:DOCUMENTATION:
}
\description{
  Brief summary of repository

  Brief summary of signature
}
\keyword{methods}

