% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential.r
\name{ssh_key_needs_passphrase}
\alias{ssh_key_needs_passphrase}
\title{Check if private key is passphrase protected}
\usage{
ssh_key_needs_passphrase(privatekey = "~/.ssh/id_rsa")
}
\arguments{
\item{privatekey}{The path to the private key of the
credential. Default is \code{'~/.ssh/id_rsa'}}
}
\description{
Check if private key is passphrase protected
}
\keyword{internal}
