% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgit2.r
\name{ssl_cert_locations}
\alias{ssl_cert_locations}
\title{Set the SSL certificate-authority locations}
\usage{
ssl_cert_locations(filename = NULL, path = NULL)
}
\arguments{
\item{filename}{Location of a file containing several certificates
concatenated together. Default NULL.}

\item{path}{Location of a directory holding several certificates,
one per file. Default NULL.}
}
\value{
invisible(NULL)
}
\description{
Set the SSL certificate-authority locations
}
\note{
Either parameter may be 'NULL', but not both.
}
\keyword{methods}

