% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{stash_drop}
\alias{stash_drop}
\alias{stash_drop,git_repository,integer-method}
\alias{stash_drop,git_repository,numeric-method}
\alias{stash_drop,git_stash,missing-method}
\title{Drop stash}
\usage{
stash_drop(object, index)

\S4method{stash_drop}{git_repository,integer}(object, index)

\S4method{stash_drop}{git_repository,numeric}(object, index)

\S4method{stash_drop}{git_stash,missing}(object)
}
\arguments{
\item{object}{The stash \code{object} to drop or a zero-based
integer to the stash to drop. The last stash has index 0.}

\item{index}{Zero based index to the stash to drop.}
}
\value{
invisible NULL
}
\description{
Drop stash
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

# Configure a user
config(repo, user.name="User", user.email="user@example.org")

# Create a file, add and commit
writeLines("Hello world!", file.path(path, "test.txt"))
add(repo, 'test.txt')
commit(repo, "Commit message")

# Change file
writeLines(c("Hello world!", "HELLO WORLD!"), file.path(path, "test.txt"))

# Create stash in repository
stash(repo)

# Change file
writeLines(c("Hello world!", "HeLlO wOrLd!"), file.path(path, "test.txt"))

# Create stash in repository
stash(repo)

# View stashes
stash_list(repo)

# Drop git_stash object in repository
stash_drop(stash_list(repo)[[1]])

## Drop stash using an index to stash
stash_drop(repo, 0)

# View stashes
stash_list(repo)
}
}
\keyword{methods}

