% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{status}
\alias{status}
\alias{status,git_repository-method}
\title{Status}
\usage{
status(repo, staged = TRUE, unstaged = TRUE, untracked = TRUE,
  ignored = FALSE, verbose = TRUE)

\S4method{status}{git_repository}(repo, staged = TRUE, unstaged = TRUE,
  untracked = TRUE, ignored = FALSE, verbose = TRUE)
}
\arguments{
\item{repo}{The \code{git_repository} to get status from.}

\item{staged}{Include staged files. Default TRUE.}

\item{unstaged}{Include unstaged files. Default TRUE.}

\item{untracked}{Include untracked files. Default TRUE.}

\item{ignored}{Include ignored files. Default FALSE.}

\item{verbose}{Display status. Default TRUE.}
}
\value{
invisible(list) with repository status
}
\description{
Display state of the repository working directory and the staging
area.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

status(repo)
}
}
\keyword{methods}

