% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,git_diff-method}
\alias{summary,git_diff-method}
\alias{summary,git_diff-methods}
\title{Show the summary of a diff}
\usage{
\S4method{summary}{git_diff}(object, ...)
}
\arguments{
\item{object}{The diff \code{object}.}
}
\description{
Show the summary of a diff
}
\keyword{methods}

