% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{summary,git_repository-method}
\alias{summary,git_repository-method}
\alias{summary,git_repository-methods}
\title{Summary of repository}
\usage{
\S4method{summary}{git_repository}(object, ...)
}
\arguments{
\item{object}{The repository \code{object}}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of repository
}
\keyword{methods}

