% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\docType{methods}
\name{summary,git_tag-method}
\alias{summary,git_tag-method}
\alias{summary,git_tag-methods}
\title{Summary of a tag}
\usage{
\S4method{summary}{git_tag}(object, ...)
}
\arguments{
\item{object}{The tag \code{object}}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of a tag
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user
config(repo, user.name="Alice", user.email="alice@example.org")

## Commit a text file
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Create tag
tag(repo, "Tagname", "Tag message")

## Summary of tag
summary(tags(repo)[[1]])
}
}
\keyword{methods}
