\docType{class}
\name{tag-class}
\alias{show,tag-method}
\alias{tag-class}
\title{Class \code{"tag"}}
\description{
  S4 class to handle a git tag, contains
  \code{\link{reference}}.
}
\section{Methods}{
  \describe{ \item{show}{\code{signature(object = "tag")}}
  }
}
\keyword{classes}
\keyword{methods}

