% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tag}
\alias{tag}
\alias{tag,git_repository-method}
\title{Tag}
\usage{
tag(object, name, message, tagger = default_signature(object))

\S4method{tag}{git_repository}(object, name, message,
  tagger = default_signature(object))
}
\arguments{
\item{object}{The repository \code{object}.}

\item{name}{Name for the tag.}

\item{message}{The tag message.}

\item{tagger}{The tagger (author) of the tag}
}
\value{
invisible(\code{git_tag}) object
}
\description{
Tag
}
\keyword{methods}

