% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,git_tree,integer,missing-method}
\alias{[,git_tree,integer,missing-method}
\title{Extract object from tree}
\usage{
\S4method{[}{git_tree,integer,missing}(x, i)
}
\arguments{
\item{x}{The tree \code{object}}

\item{i}{The index (integer or logical) of the tree object to
extract. If negative values, all elements except those indicated
are selected. A character vector to match against the names of
objects to extract.}
}
\value{
Git object
}
\description{
Lookup a tree entry by its position in the tree
}
\keyword{methods}

