% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{tree}
\alias{tree}
\alias{tree,git_commit-method}
\title{Tree}
\usage{
tree(object)

\S4method{tree}{git_commit}(object)
}
\arguments{
\item{object}{the \code{commit} object}
}
\value{
A S4 class git_tree object
}
\description{
Get the tree pointed to by a commit.
}
\keyword{methods}

