% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{workdir}
\alias{workdir}
\title{Workdir of repository}
\usage{
workdir(repo = ".")
}
\arguments{
\item{repo}{a path to a repository or a
\code{\linkS4class{git_repository}} object. Default is '.'}
}
\value{
Character vector with the path of the workdir. If the
repository is bare, \code{NULL} will be returned.
}
\description{
Workdir of repository
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)

## Get the path of the workdir for repository
workdir(repo)
}
}
