% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{add}
\alias{add}
\alias{add,git_repository-method}
\title{Add file(s) to index}
\usage{
add(object, path)

\S4method{add}{git_repository}(object, path)
}
\arguments{
\item{object}{The repository \code{object}.}

\item{path}{character vector with filenames to add. The path must
be relative to the repository's working filder.}
}
\value{
invisible(NULL)
}
\description{
Add file(s) to index
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Add file repository
add(repo, "file-to-add")
}
}
\keyword{methods}

