% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ahead_behind}
\alias{ahead_behind}
\alias{ahead_behind,git_commit,git_commit-method}
\title{Ahead Behind}
\usage{
ahead_behind(local, upstream)

\S4method{ahead_behind}{git_commit,git_commit}(local, upstream)
}
\arguments{
\item{local}{a S4 class git_commit \code{object}.}

\item{upstream}{a S4 class git_commit \code{object}.}
}
\value{
An integer vector of length 2 with number of commits that
the upstream commit is ahead and behind the local commit
}
\description{
Count the number of unique commits between two commit objects.
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)
config(repo, user.name="Developer", user.email="developer@example.org")

## Create a file, add and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           con = file.path(path, "test.txt"))
add(repo, "test.txt")
commit_1 <- commit(repo, "Commit message 1")

# Change file and commit
writeLines(c("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
             "eiusmod tempor incididunt ut labore et dolore magna aliqua."),
             con = file.path(path, "test.txt"))
add(repo, "test.txt")
commit_2 <- commit(repo, "Commit message 2")

ahead_behind(commit_1, commit_2)
}
}
\keyword{methods}

