% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_delete}
\alias{branch_delete}
\alias{branch_delete,git_branch-method}
\title{Delete a branch}
\usage{
branch_delete(branch)

\S4method{branch_delete}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
invisible NULL
}
\description{
Delete a branch
}
\keyword{methods}

