% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{branch_remote_name}
\alias{branch_remote_name}
\alias{branch_remote_name,git_branch-method}
\title{Remote name of a branch}
\usage{
branch_remote_name(branch)

\S4method{branch_remote_name}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
character string with remote name
}
\description{
The name of remote that the remote tracking branch belongs to
}
\keyword{methods}

