% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_rename}
\alias{branch_rename}
\alias{branch_rename,git_branch-method}
\title{Rename a branch}
\usage{
branch_rename(branch, name, force = FALSE, message = NULL, who = NULL)

\S4method{branch_rename}{git_branch}(branch, name, force = FALSE,
  message = NULL, who = NULL)
}
\arguments{
\item{branch}{Branch to rename}

\item{name}{The new name for the branch}

\item{force}{Overwrite existing branch. Default is FALSE}

\item{signature}{The identity that will be used to populate the
reflog entry. Default is NULL, which gives the default signature.}

\item{message}{The one line long message to the reflog. If NULL,
the default value is appended}
}
\value{
invisible renamed S4 class git_branch
}
\description{
Rename a branch
}
\keyword{methods}

