% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_target}
\alias{branch_target}
\alias{branch_target,branch-method}
\title{Get hex pointed to by a branch}
\usage{
branch_target(branch)

\S4method{branch_target}{branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
hex or NA if not a direct reference
}
\description{
Get hex pointed to by a branch
}
\keyword{methods}

