% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branch_target}
\alias{branch_target}
\alias{branch_target,git_branch-method}
\title{Get target (sha) pointed to by a branch}
\usage{
branch_target(branch)

\S4method{branch_target}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch}
}
\value{
sha or NA if not a direct reference
}
\description{
Get target (sha) pointed to by a branch
}
\keyword{methods}

