\docType{methods}
\name{branches-methods}
\alias{branches}
\alias{branches-methods}
\title{Branches}
\usage{
branches(object, flags = c("ALL", "LOCAL", "REMOTE"))
}
\arguments{
  \item{object}{:TODO:DOCUMENTATION:}

  \item{flags}{:TODO:DOCUMENTATION:}
}
\value{
:TODO:DOCUMENTATION:
}
\description{
Branches
}
\keyword{methods}

