\docType{methods}
\name{branches-methods}
\alias{branches}
\alias{branches-methods}
\alias{branches,repository-method}
\title{Branches}
\arguments{
  \item{object}{:TODO:DOCUMENTATION:}

  \item{flags}{:TODO:DOCUMENTATION:}
}
\value{
  :TODO:DOCUMENTATION:
}
\description{
  Branches
}
\keyword{methods}

