% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{branches}
\alias{branches}
\alias{branches,git_repository-method}
\title{Branches}
\usage{
branches(repo, flags = c("all", "local", "remote"))

\S4method{branches}{git_repository}(repo, flags = c("all", "local", "remote"))
}
\arguments{
\item{repo}{The repository \code{object}}

\item{flags}{Filtering flags for the branch listing. Valid values
are 'all', 'local' or 'remote'}
}
\value{
list of branches in repository
}
\description{
List branches in repository
}
\keyword{methods}

