% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{branches}
\alias{branches}
\alias{branches,git_repository-method}
\title{Branches}
\usage{
branches(object, flags = c("ALL", "LOCAL", "REMOTE"))

\S4method{branches}{git_repository}(object, flags = c("ALL", "LOCAL",
  "REMOTE"))
}
\arguments{
\item{object}{The repository}

\item{flags}{Filtering flags for the branch listing. Valid values
are 'ALL', 'LOCAL' or 'REMOTE'}
}
\value{
list of branches in repository
}
\description{
List branches in repository
}
\keyword{methods}

