% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{checkout}
\alias{checkout}
\alias{checkout,git_branch-method}
\alias{checkout,git_commit-method}
\alias{checkout,git_repository-method}
\alias{checkout,git_tag-method}
\title{Checkout}
\usage{
checkout(object, ...)

\S4method{checkout}{git_repository}(object, branch, create = FALSE,
  force = FALSE)

\S4method{checkout}{git_branch}(object, force = FALSE)

\S4method{checkout}{git_commit}(object, force = FALSE)

\S4method{checkout}{git_tag}(object, force = FALSE)
}
\arguments{
\item{object}{a repository, commit, tag or tree which content will
be used to update the working directory.}

\item{...}{Additional arguments affecting the checkout}

\item{branch}{If object is a repository, the name of the branch to
check out.}

\item{create}{If object is a repository, then branch is created if
doesn't exist.}

\item{force}{If TRUE, then make working directory match
target. This will throw away local changes. Default is FALSE.}
}
\value{
invisible NULL
}
\description{
Update files in the index and working tree to match the content of
the tree pointed at by the treeish object (commit, tag or tree).
Checkout using the default GIT_CHECKOUT_SAFE_CREATE strategy
(force = FALSE) or GIT_CHECKOUT_FORCE (force = TRUE).
}
\keyword{methods}

