% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{checkout}
\alias{checkout}
\alias{checkout,git_commit-method}
\alias{checkout,git_tag-method}
\alias{checkout,git_tree-method}
\title{Checkout}
\usage{
checkout(object)

\S4method{checkout}{git_commit}(object)

\S4method{checkout}{git_tag}(object)

\S4method{checkout}{git_tree}(object)
}
\arguments{
\item{object}{a commit, tag or tree which content will be used to
update the working directory.}
}
\value{
invisible NULL
}
\description{
Update files in the index and working tree to match the content of
the tree pointed at by the treeish object (commit, tag or tree).
}
\keyword{methods}

