% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkout.r
\name{checkout_reflog_msg}
\alias{checkout_reflog_msg}
\title{Internal function to generate checkout reflog message}
\usage{
checkout_reflog_msg(object, ref_log_target)
}
\arguments{
\item{object}{The object to checkout}

\item{ref_log_target}{The target in the reflog message
in the reflog}
}
\description{
Internal function to generate checkout reflog message
}
\keyword{internal}

