% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clone}
\alias{clone}
\title{Clone a remote repository}
\usage{
clone(url, local_path, progress = TRUE)
}
\arguments{
\item{url}{the remote repository to clone}

\item{local_path}{local directory to clone to}

\item{progress}{show progress}
}
\value{
A S4 \code{git_repository} object
}
\description{
Clone a remote repository
}
\examples{
\dontrun{
## Clone a remote repository
repo <- clone("https://github.com/ropensci/git2r", "path/to/git2r")
}
}
\keyword{methods}

