% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tree.r
\docType{methods}
\name{coerce-git_tree-data.frame-method}
\alias{coerce,git_tree,data.frame-method}
\alias{coerce-git_tree-data.frame-method}
\title{Coerce entries in a git_tree to a \code{data.frame}}
\arguments{
\item{from}{The tree \code{object}}
}
\value{
\code{data.frame}
}
\description{
The entries in a tree are coerced to a \code{data.frame}
}
\details{
The \code{data.frame} have the following columns:
\describe{

\item{filemode}{
    The UNIX file attributes of a tree entry
  }

\item{type}{
    String representation of the tree entry type
  }

\item{sha}{
    The sha of a tree entry
  }

\item{name}{
    The filename of a tree entry
  }

}
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Coerce tree to a data.frame
df <- as(tree(commits(repo)[[1]]), "data.frame")

str(df)
}
}
\keyword{methods}

