% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{commit}
\alias{commit}
\alias{commit,git_repository-method}
\title{Commit}
\usage{
commit(object, message = NULL, reference = "HEAD",
  author = default_signature(object), committer = default_signature(object))

\S4method{commit}{git_repository}(object, message = NULL,
  reference = "HEAD", author = default_signature(object),
  committer = default_signature(object))
}
\arguments{
\item{object}{The repository \code{object}.}

\item{message}{The commit message.}

\item{reference}{Name of the reference that will be updated to
point to this commit.}

\item{author}{Signature with author and author time of commit.}

\item{committer}{Signature with committer and commit time of commit.}
}
\value{
\code{git_commit} object
}
\description{
Commit
}
\keyword{methods}

