% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{commit-methods}
\alias{commit}
\alias{commit-methods}
\title{Commit}
\usage{
commit(object, message = NULL, reference = "HEAD",
  author = default_signature(object), committer = default_signature(object))
}
\arguments{
\item{object}{The repository \code{object}.}

\item{reference}{:TODO:DOCUMENTATION:}

\item{author}{:TODO:DOCUMENTATION:}

\item{committer}{:TODO:DOCUMENTATION:}

\item{message}{:TODO:DOCUMENTATION:}
}
\value{
\code{git_commit} object
}
\description{
Commit
}
\keyword{methods}

