\docType{methods}
\name{commit-methods}
\alias{commit}
\alias{commit,git_repository-method}
\alias{commit-methods}
\title{Commit}
\arguments{
  \item{object}{The repository \code{object}.}

  \item{reference}{:TODO:DOCUMENTATION:}

  \item{author}{:TODO:DOCUMENTATION:}

  \item{committer}{:TODO:DOCUMENTATION:}

  \item{message}{:TODO:DOCUMENTATION:}

  \item{tree}{:TODO:DOCUMENTATION:}

  \item{parents}{:TODO:DOCUMENTATION:}
}
\value{
  invisible(NULL)
}
\description{
  Commit
}
\keyword{methods}

