% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{commits}
\alias{commits}
\alias{commits,character-method}
\alias{commits,git_repository-method}
\title{Commits}
\usage{
commits(repo, topological = TRUE, time = TRUE, reverse = FALSE)

\S4method{commits}{character}(repo, topological = TRUE, time = TRUE,
  reverse = FALSE)

\S4method{commits}{git_repository}(repo, topological = TRUE, time = TRUE,
  reverse = FALSE)
}
\arguments{
\item{repo}{The repository \code{object}.}

\item{topological}{Sort the commits in topological order (parents
before children); can be combined with time sorting. Default is
TRUE.}

\item{time}{Sort the commits by commit time; Can be combined with
topological sorting. Default is TRUE.}

\item{reverse}{Sort the commits in reverse order; can be combined
with topological and/or time sorting. Default is FALSE.}
}
\value{
list of commits in repository
}
\description{
Commits
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## List commits in repository
commits(repo)

## List commits direct from path
commits("path/to/git2r")
}
}
\keyword{methods}

