% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{commits}
\alias{commits}
\alias{commits,character-method}
\alias{commits,git_repository-method}
\title{Commits}
\usage{
commits(object)

\S4method{commits}{character}(object)

\S4method{commits}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object}.}
}
\value{
list of commits in repository
}
\description{
Commits
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## List commits in repository
commits(repo)

## List commits direct from path
commits("path/to/git2r")
}
}
\keyword{methods}

