% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{config}
\alias{config}
\alias{config,git_repository-method}
\title{Config}
\usage{
config(repo, user.name, user.email)

\S4method{config}{git_repository}(repo, user.name, user.email)
}
\arguments{
\item{repo}{the \code{repo} to configure}

\item{user.name}{the user name}

\item{user.email}{the e-mail address}
}
\value{
A \code{list} with the configuration
}
\description{
Config file management.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository('path/to/git2r')

config(repo, user.name='Stefan Widgren', user.email='stefan.widgren@gmail.com')
}
}
\keyword{methods}

