% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{content}
\alias{content}
\alias{content,git_blob-method}
\title{Content of blob}
\usage{
content(blob, lines = TRUE, n = -1)

\S4method{content}{git_blob}(blob, lines = TRUE, n = -1)
}
\arguments{
\item{object}{The blob \code{object}.}

\item{lines}{Read blob content as text lines. Default TRUE.}

\item{n}{integer. If lines is TRUE, the (maximal) number of lines
to read. Negative values indicate that one should read up to the
end of input from blob content.}
}
\value{
The content of the blob
}
\description{
Content of blob
}
\keyword{methods}

