% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{cred_user_pass-class}
\alias{cred_user_pass-class}
\title{S4 class to handle plain-text username and password
credential object}
\description{
Class \code{"cred_user_pass"}
}
\section{Slots}{

\describe{
  \item{username}{
    The username of the credential
  }
  \item{password}{
    The password of the credential
  }
}
}
\keyword{classes}
\keyword{methods}

