% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/credential.r
\name{cred_user_pass}
\alias{cred_user_pass}
\title{Create a new plain-text username and password credential object}
\usage{
cred_user_pass(username, password)
}
\arguments{
\item{username}{The username of the credential}

\item{password}{The password of the credential}
}
\value{
A S4 \code{cred_user_pass} object
}
\description{
Create a new plain-text username and password credential object
}
\examples{
\dontrun{
## Create a plain-text username and password credential object
cred_user_pass("Random Developer", "SecretPassword")
}
}
\seealso{
Other git.credential.functions: \code{\link{cred_env}};
  \code{\link{cred_ssh_key}}; \code{\link{cred_token}}
}
\keyword{methods}

