% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{default_signature}
\alias{default_signature}
\alias{default_signature,git_repository-method}
\title{Get the signature}
\usage{
default_signature(object)

\S4method{default_signature}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object} to check signature}
}
\value{
Character vector with signature
}
\description{
Get the signature according to the repository's configuration
}
\keyword{methods}

