% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{default_signature}
\alias{default_signature}
\alias{default_signature,git_repository-method}
\title{Get the signature}
\usage{
default_signature(repo)

\S4method{default_signature}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository \code{object} to check signature}
}
\value{
S4 class git_signature
}
\description{
Get the signature according to the repository's configuration
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Get the default signature
default_signature(repo)
}
}
\keyword{methods}

