% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{default_signature}
\alias{default_signature}
\title{Get the signature}
\usage{
default_signature(repo = NULL)
}
\arguments{
\item{repo}{a \code{\linkS4class{git_repository}} object, or a
path to a repository, or \code{NULL}.  If the \code{repo}
argument is \code{NULL}, the repository is searched for with
\code{\link{discover_repository}} in the current working
directory.}
}
\value{
S4 class git_signature
}
\description{
Get the signature according to the repository's configuration
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user
config(repo, user.name="Alice", user.email="alice@example.org")

## Get the default signature
default_signature(repo)

## Change user
config(repo, user.name="Bob", user.email="bob@example.org")

## Get the default signature
default_signature(repo)
}
}
