% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{diff}
\alias{diff}
\alias{diff,git_repository-method}
\alias{diff,git_tree-method}
\title{Changes between commits, trees, working tree, etc.}
\usage{
diff(object, ...)

\S4method{diff}{git_repository}(object, index = FALSE)

\S4method{diff}{git_tree}(object, new_tree = NULL, index = FALSE)
}
\arguments{
\item{object}{A git_repository object}

\item{index}{Flag, whether to compare the index to HEAD. If
FALSE (the default), then the working tree is compared to
the index.}

\item{object}{The old git_tree object to compare to}

\item{new_tree}{The new git_tree object to compare, or NULL.
If NULL, then we use the working directory or the index (see
the \code{index} argument).}

\item{index}{Whether to use the working directory (by default),
or the index (if set to TRUE) in the comparison to
\code{object}.}
}
\value{
A git_diff object
}
\description{
Changes between commits, trees, working tree, etc.
}
\keyword{methods}

