% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{diff}
\alias{diff}
\alias{diff,git_repository-method}
\alias{diff,git_tree-method}
\title{Changes between commits, trees, working tree, etc.}
\usage{
diff(object, ...)

\S4method{diff}{git_repository}(object, index = FALSE, filename = NULL)

\S4method{diff}{git_tree}(object, new_tree = NULL, index = FALSE,
  filename = NULL)
}
\arguments{
\item{object}{A \code{git_repository} object or the old
\code{git_tree} object to compare to.}

\item{index}{\describe{
  \item{\emph{When object equals a git_repository}}{
    Whether to compare the index to HEAD. If FALSE (the default),
    then the working tree is compared to the index.
  }
  \item{\emph{When object equals a git_tree}}{
    Whether to use the working directory (by default), or the index
    (if set to TRUE) in the comparison to \code{object}.
  }
}}

\item{filename}{Determines where to write the diff. If filename is
NULL (the default), then the diff is written to a S4 class
\code{\linkS4class{git_diff}} object. If filename equals
\code{character(0)} or \code{NA_character_}, then the diff is
written to a character vector. If filename is a character vector of
length one with non-NA value, the diff is written to a file with
name filename (the file is overwritten if it exists).}

\item{...}{Additional arguments affecting the diff produced}

\item{new_tree}{The new git_tree object to compare, or NULL.
If NULL, then we use the working directory or the index (see
the \code{index} argument).}
}
\value{
A git_diff object if filename equals NULL. A character
vector if filename equals \code{character(0)} or
\code{NA_character_}. Oterwise NULL.
}
\description{
Changes between commits, trees, working tree, etc.
}
\keyword{methods}

