% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fetch}
\alias{fetch}
\alias{fetch,git_repository-method}
\title{Fetch new data and update tips}
\usage{
fetch(repo, name, credentials = NULL)

\S4method{fetch}{git_repository}(repo, name, credentials = NULL)
}
\arguments{
\item{repo}{the repository}

\item{name}{the remote's name}

\item{credentials}{The credentials for remote repository
access. Default is NULL.}
}
\value{
invisible \code{\linkS4class{git_transfer_progress}} object
}
\description{
Fetch new data and update tips
}
\keyword{methods}

