% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.r
\docType{methods}
\name{fetch}
\alias{fetch}
\alias{fetch,git_repository-method}
\title{Fetch new data and update tips}
\usage{
fetch(repo, name, credentials = NULL)

\S4method{fetch}{git_repository}(repo, name, credentials = NULL)
}
\arguments{
\item{repo}{the repository}

\item{name}{the remote's name}

\item{credentials}{The credentials for remote repository
access. Default is NULL. To use and query an ssh-agent for the ssh
key credentials, let this parameter be NULL (the default).}
}
\value{
invisible \code{\linkS4class{git_transfer_progress}} object
}
\description{
Fetch new data and update tips
}
\examples{
\dontrun{
## Initialize three temporary repositories
path_bare <- tempfile(pattern="git2r-")
path_repo_1 <- tempfile(pattern="git2r-")
path_repo_2 <- tempfile(pattern="git2r-")

dir.create(path_bare)
dir.create(path_repo_1)
dir.create(path_repo_2)

bare_repo <- init(path_bare, bare = TRUE)
repo_1 <- clone(path_bare, path_repo_1)
repo_2 <- clone(path_bare, path_repo_2)

config(repo_1, user.name="Alice", user.email="alice@example.org")
config(repo_2, user.name="Bob", user.email="bob@example.org")

## Add changes to repo 1
writeLines("Lorem ipsum dolor sit amet",
           con = file.path(path_repo_1, "example.txt"))
add(repo_1, "example.txt")
commit(repo_1, "Commit message")

## Push changes from repo 1 to origin (bare_repo)
push(repo_1, "origin", "refs/heads/master")

## Fetch changes from origin (bare_repo) to repo 2
fetch(repo_2, "origin")

## List updated heads
fetch_heads(repo_2)
}
}
\keyword{methods}

