% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refspec.r
\name{get_refspec}
\alias{get_refspec}
\title{Create push refspec from arguments}
\usage{
get_refspec(repo = NULL, remote = NULL, spec = NULL)
}
\arguments{
\item{repo}{S4 class \code{git_repository}}

\item{name}{The remote's name. Default is NULL.}

\item{refspec}{The refspec to be pushed. Default is NULL.}
}
\value{
Character vector with refspec
}
\description{
Create push refspec from arguments
}
\keyword{internal}

