% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git2r.R
\docType{package}
\name{git2r}
\alias{git2r}
\alias{git2r-package}
\title{git2r: R bindings to the libgit2 library}
\description{
git2r: R bindings to the libgit2 library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/git2r/}
  \item \url{https://github.com/ropensci/git2r}
  \item Report bugs at \url{https://github.com/ropensci/git2r/issues}
}

}
\author{
\strong{Maintainer}: Stefan Widgren \email{stefan.widgren@gmail.com} (\href{https://orcid.org/0000-0001-5745-2284}{ORCID})

Other contributors:
\itemize{
  \item Gabor Csardi [contributor]
  \item Gregory Jefferis [contributor]
  \item Jennifer Bryan [contributor]
  \item Jeroen Ooms [contributor]
  \item Jim Hester [contributor]
  \item John Blischak [contributor]
  \item Karthik Ram [contributor]
  \item Peter Carbonetto [contributor]
  \item Scott Chamberlain [contributor]
  \item Thomas Rosendal [contributor]
}

}
