% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_blame-class}
\alias{git_blame-class}
\title{S4 class to handle a git blame for a single file}
\description{
Class \code{"git_blame"}
}
\section{Slots}{

\describe{
  \item{path}{
    The path to the file of the blame
  }
  \item{hunks}{
    List of blame hunks
  }
  \item{repo}{
    The S4 class git_repository that contains the file
  }
}
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Get blame for file in repository
blame(repo, ".gitignore")
}
}
\keyword{classes}
\keyword{methods}

