% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{git_blob-class}
\alias{git_blob-class}
\title{S4 class to handle a git blob}
\description{
Class \code{"git_blob"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{path}{
    Path to the git repository that contains the blob
  }
}
}
\keyword{classes}
\keyword{methods}

