% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_blob-class}
\alias{git_blob-class}
\title{S4 class to handle a git blob}
\description{
Class \code{"git_blob"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{repo}{
    The S4 class git_repository that contains the blob
  }
}
}
\keyword{classes}
\keyword{methods}

