% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_branch-class}
\alias{git_branch-class}
\title{S4_class_to_handle_a_git_branch}
\description{
Class \code{git_branch}
}
\section{Slots}{

\describe{
  \item{remote}{
    The name of remote that the remote tracking branch belongs to
  }
  \item{url}{
    The remote's url
  }
  \item{head}{
    TRUE if the current local branch is pointed at by HEAD
  }
}
}

\section{Methods}{

\describe{
  \item{is.head}{\code{signature(object = "git_branch")}}
  \item{is.local}{\code{signature(object = "git_branch")}}
  \item{show}{\code{signature(object = "git_branch")}}
}
}
\keyword{classes}
\keyword{methods}

