% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_branch-class}
\alias{git_branch-class}
\title{S4_class_to_handle_a_git_branch}
\description{
Class \code{git_branch}
}
\section{Slots}{

\describe{
  \item{name}{
    Name of the branch.
  }
  \item{type}{
    Type of the branch, either 1 (local) or 2 (remote).
  }
  \item{repo}{
    The S4 class git_repository that contains the branch
  }
}
}

\section{Methods}{

\describe{
  \item{is_head}{\code{signature(object = "git_branch")}}
  \item{is_local}{\code{signature(object = "git_branch")}}
  \item{show}{\code{signature(object = "git_branch")}}
}
}
\keyword{classes}
\keyword{methods}

