\docType{class}
\name{git_branch-class}
\alias{git_branch-class}
\alias{show,git_branch-method}
\title{S4_class_to_handle_a_git_branch}
\description{
S4_class_to_handle_a_git_branch
}
\section{Slots}{
  \describe{ \item{remote}{ :TODO:DOCUMENTATION: }
  \item{url}{ :TODO:DOCUMENTATION: } \item{head}{
  :TODO:DOCUMENTATION: } }
}

\section{Methods}{
  \describe{ \item{is.head}{\code{signature(object =
  "git_branch")}} \item{is.local}{\code{signature(object =
  "git_branch")}} \item{show}{\code{signature(object =
  "git_branch")}} }
}
\keyword{classes}
\keyword{methods}

