% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_commit-class}
\alias{git_commit-class}
\title{S4 class to handle a git commit.}
\description{
Class \code{"git_commit"}
}
\section{Slots}{

\describe{
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1
  }
  \item{author}{
    An author signature
  }
  \item{committer}{
    The committer signature
  }
  \item{summary}{
    The short "summary" of a git commit message, comprising the first
    paragraph of the message with whitespace trimmed and squashed.
  }
  \item{message}{
    The message of a commit
  }
  \item{repo}{
    The S4 class git_repository that contains the commit
  }
}
}

\section{Methods}{

\describe{
  \item{show}{\code{signature(object = "git_commit")}}
}
}
\keyword{classes}
\keyword{methods}

