% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_diff-class}
\alias{git_diff-class}
\title{Git diff}
\description{
Git diff
}
\section{Slots}{

\describe{
  \item{old}{
    Represents the 'from' side of the diff.
  }
  \item{new}{
    Represents the 'to' side of the diff.
  }
  \item{files}{
    List of \code{\linkS4class{git_diff_file}} objects.
  }
}
}

\keyword{classes}
