% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{git_diff_file-class}
\alias{git_diff_file-class}
\title{Git diff file}
\description{
Git diff file
}
\section{Slots}{

\describe{
  \item{old_file}{
    Filename for old side of diff.
  }
  \item{new_file}{
    Filename for new side of diff.
  }
  \item{hunks}{
    List of \code{\linkS4class{git_diff_hunk}} objects.
  }
}
}

\keyword{classes}
