% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{git_diff_hunk-class}
\alias{git_diff_hunk-class}
\title{Git diff hunk}
\description{
Git diff hunk
}
\section{Slots}{

\describe{
  \item{old_start}{
    Starting line number in old_file.
  }
  \item{old_lines}{
    Number of lines in old_file.
  }
  \item{new_start}{
    Starting line number in new_file.
  }
  \item{new_lines}{
    Number of lines in new_file.
  }
  \item{header}{
    Header text.
  }
  \item{lines}{
    List with \code{\linkS4class{git_diff_line}} objects.
  }
}
}

\keyword{classes}
