% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_diff_line-class}
\alias{git_diff_line-class}
\title{Git diff line}
\description{
Git diff line
}
\section{Slots}{

\describe{
  \item{origin}{
    Line origin constants.
  }
  \item{old_lineno}{
    Line number in old file or -1 for added line.
  }
  \item{new_lineno}{
    Line number in new file or -1 for deleted line.
  }
  \item{num_lines}{
    Number of newline characters in content.
  }
  \item{content}{
    The diff text.
  }
}
}
\keyword{classes}

