% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_note-class}
\alias{git_note-class}
\title{S4 class to handle a git note}
\description{
Class \code{git_note}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{note}{
    The note message
  }
  \item{repo}{
    The S4 class git_repository that contains the note
  }
}
}
\keyword{classes}
\keyword{methods}

