% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_note-class}
\alias{git_note-class}
\title{S4 class to handle a git note}
\description{
Class \code{git_note}
}
\section{Slots}{

\describe{
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1 of the blob
    containing the message
  }
  \item{annotated}{
    The 40 character hexadecimal string of the SHA-1 of the git
    object being annotated
  }
  \item{message}{
    The note message
  }
  \item{refname}{
    Name of the reference
  }
  \item{repo}{
    The S4 class git_repository that contains the note
  }
}
}
\keyword{classes}
\keyword{methods}

