\docType{class}
\name{git_repository-class}
\alias{git_repository-class}
\title{Class \code{"git_repository"}}
\description{
S4 class to handle a git repository
}
\section{Slots}{
  \describe{ \item{path}{ Path to a git repository } }
}

\section{Methods}{
  \describe{ \item{is.bare}{\code{signature(object =
  "git_repository")}}
  \item{is.empty}{\code{signature(object =
  "git_repository")}} }
}
\keyword{classes}
\keyword{methods}

