% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_repository-class}
\alias{git_repository-class}
\title{S4 class to handle a git repository}
\description{
Class \code{"git_repository"}
}
\section{Slots}{

\describe{
  \item{path}{
    Path to a git repository
  }
}
}

\section{Methods}{

\describe{
  \item{is_bare}{\code{signature(object = "git_repository")}}
  \item{is_empty}{\code{signature(object = "git_repository")}}
}
}
\keyword{classes}
\keyword{methods}

