\docType{class}
\name{git_signature-class}
\alias{git_signature-class}
\alias{show,git_signature-method}
\title{Class \code{"git_signature"}}
\description{
S4 class to handle a git signature
}
\section{Slots}{
  \describe{ \item{name}{ The full name of the author.  }
  \item{email}{ Email of the author.  } \item{when}{ Time
  when the action happened.  } }
}

\section{Methods}{
  \describe{ \item{show}{\code{signature(object =
  "git_signature")}} }
}
\keyword{classes}
\keyword{methods}

