% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_signature-class}
\alias{git_signature-class}
\title{S4 class to handle a git signature}
\description{
Class \code{"git_signature"}
}
\section{Slots}{

\describe{
  \item{name}{
    The full name of the author.
  }
  \item{email}{
    Email of the author.
  }
  \item{when}{
    Time when the action happened.
  }
}
}
\keyword{classes}
\keyword{methods}

