\docType{class}
\name{git_tag-class}
\alias{git_tag-class}
\alias{show,git_tag-method}
\title{Class \code{"git_tag"}}
\description{
  S4 class to handle a git tag, contains
  \code{git_reference}.
}
\section{Slots}{
  \describe{ \item{sig}{ An action signature } }
}

\section{Methods}{
  \describe{ \item{show}{\code{signature(object =
  "git_tag")}} }
}
\keyword{classes}
\keyword{methods}

