% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_tag-class}
\alias{git_tag-class}
\title{S4 class to handle a git tag}
\description{
Class \code{"git_tag"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{message}{
    The message of the tag
  }
  \item{name}{
    The name of the tag
  }
  \item{tagger}{
    The tagger (author) of the tag
  }
  \item{target}{
    The target of the tag
  }
  \item{repo}{
    The S4 class git_repository that contains the tag
  }
}
}
\keyword{classes}
\keyword{methods}

