% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_transfer_progress-class}
\alias{git_transfer_progress-class}
\title{Class \code{"git_transfer_progress"}}
\description{
Statistics from the fetch operation.
}
\section{Slots}{

\describe{
  \item{total_objects}{
    Number of objects in the packfile being downloaded
  }
  \item{indexed_objects}{
    Received objects that have been hashed
  }
  \item{received_objects}{
    Objects which have been downloaded
  }
  \item{total_deltas}{
    :TODO:DOCUMENTATION:
  }
  \item{indexed_deltas}{
    :TODO:DOCUMENTATION:
  }
  \item{local_objects}{
    Locally-available objects that have been injected in order to
    fix a thin pack
  }
  \item{received_bytes}{
    Size of the packfile received up to now
  }
}
}
\keyword{classes}
\keyword{methods}

