% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_tree-class}
\alias{git_tree-class}
\title{S4 class to handle a git tree}
\description{
Class \code{"git_tree"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{filemode}{
    :TODO:DOCUMENTATION:
  }
  \item{type}{
    :TODO:DOCUMENTATION:
  }
  \item{id}{
    :TODO:DOCUMENTATION:
  }
  \item{name}{
    :TODO:DOCUMENTATION:
  }
  \item{repo}{
    The S4 class git_repository that contains the commit
  }
}
}
\keyword{classes}
\keyword{methods}

