% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{git_tree-class}
\alias{git_tree-class}
\title{S4 class to handle a git tree}
\description{
Class \code{"git_tree"}
}
\section{Slots}{

\describe{
  \item{sha}{
    The 40 character hexadecimal string of the SHA-1
  }
  \item{filemode}{
    The UNIX file attributes of a tree entry
  }
  \item{type}{
    String representation of the tree entry type
  }
  \item{id}{
    The sha id of a tree entry
  }
  \item{name}{
    The filename of a tree entry
  }
  \item{repo}{
    The S4 class git_repository that contains the commit
  }
}
}

\keyword{classes}
\keyword{methods}
