% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{git_tree-class}
\alias{git_tree-class}
\title{S4 class to handle a git tree}
\description{
Class \code{"git_tree"}
}
\section{Slots}{

\describe{
  \item{hex}{
    40 char hexadecimal string
  }
  \item{filemode}{
    The UNIX file attributes of a tree entry
  }
  \item{type}{
    String representation of the tree entry type
  }
  \item{id}{
    The hex id of a tree entry
  }
  \item{name}{
    The filename of a tree entry
  }
  \item{repo}{
    The S4 class git_repository that contains the commit
  }
}
}
\keyword{classes}
\keyword{methods}

