% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{head,git_repository-method}
\alias{head,git_repository-method}
\title{Get HEAD for a repo}
\usage{
\S4method{head}{git_repository}(x)
}
\arguments{
\item{x}{The repository \code{x} to check head}
}
\value{
Character vector with head
}
\description{
Get HEAD for a repo
}
\keyword{methods}

