% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{is.head}
\alias{is.head}
\alias{is.head,git_branch-method}
\title{Check if branch is head}
\usage{
is.head(object)

\S4method{is.head}{git_branch}(object)
}
\arguments{
\item{object}{The branch \code{object} to check if it's head}
}
\value{
TRUE if branch is head, else FALSE
}
\description{
Check if branch is head
}
\keyword{methods}

