\docType{methods}
\name{is.local-methods}
\alias{is.local}
\alias{is.local-methods}
\title{Check if branch is local}
\usage{
is.local(object)
}
\arguments{
  \item{object}{The branch \code{object} to check if it's
  local}
}
\value{
TRUE if branch is local, else FALSE
}
\description{
Check if branch is local
}
\keyword{methods}

