% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{is_empty}
\alias{is_empty}
\alias{is_empty,git_repository-method}
\title{Check if repository is empty}
\usage{
is_empty(repo)

\S4method{is_empty}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository to check if it's empty}
}
\value{
TRUE if empty else FALSE
}
\description{
Check if repository is empty
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Check if it's an empty repository
is_empty(repo)
}
}
\keyword{methods}

