% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branch.r
\docType{methods}
\name{is_head}
\alias{is_head}
\alias{is_head,git_branch-method}
\title{Check if branch is head}
\usage{
is_head(branch)

\S4method{is_head}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch \code{object} to check if it's head}
}
\value{
TRUE if branch is head, else FALSE
}
\description{
Check if branch is head
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user and commit a file
config(repo, user.name="Alice", user.email="alice@example.org")
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## List branches
branches(repo)

## Check that 'master' is_head
master <- branches(repo)[[1]]
is_head(master)

## Create and checkout 'dev' branch
checkout(repo, "dev", create = TRUE)

## List branches
branches(repo)

## Check that 'master' is no longer head
is_head(master)
}
}
\keyword{methods}

