% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{is_head}
\alias{is_head}
\alias{is_head,git_branch-method}
\title{Check if branch is head}
\usage{
is_head(branch)

\S4method{is_head}{git_branch}(branch)
}
\arguments{
\item{branch}{The branch \code{object} to check if it's head}
}
\value{
TRUE if branch is head, else FALSE
}
\description{
Check if branch is head
}
\keyword{methods}

