% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{last_commit}
\alias{last_commit}
\title{Last commit}
\usage{
last_commit(repo = ".")
}
\arguments{
\item{repo}{a path to a repository or a
\code{\linkS4class{git_repository}} object. Default is '.'}
}
\description{
Get last commit in the current branch.
}
\examples{
\dontrun{
## Initialize a repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Config user
config(repo, user.name="Alice", user.email="alice@example.org")

## Write to a file and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Get last commit
last_commit(repo)
last_commit(path)
}
}
