% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{length,git_blob-method}
\alias{length,git_blob-method}
\title{Size in bytes of the contents of a blob}
\usage{
\S4method{length}{git_blob}(x)
}
\arguments{
\item{x}{The blob \code{object}}
}
\value{
a non-negative integer
}
\description{
Size in bytes of the contents of a blob
}
\keyword{methods}

