% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgit2.r
\name{libgit2_version}
\alias{libgit2_version}
\title{Version of the libgit2 library}
\usage{
libgit2_version()
}
\value{
A list with major, minor and rev
}
\description{
Version of the libgit2 library that the bundled source code is
based on
}
\examples{
libgit2_version()
}
\keyword{methods}
