% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/libgit2.r
\name{libgit2_version}
\alias{libgit2_version}
\title{Version of the libgit2 library being currently used}
\usage{
libgit2_version()
}
\value{
A list with major, minor and rev
}
\description{
Version of the libgit2 library being currently used
}
\examples{
\dontrun{
libgit2_version()
}
}
\keyword{methods}

