% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{lookup}
\alias{lookup}
\alias{lookup,git_repository-method}
\title{Lookup}
\usage{
lookup(object, id)

\S4method{lookup}{git_repository}(object, id)
}
\arguments{
\item{object}{The repository \code{object}.}

\item{id}{the identity of the blob to lookup}
}
\value{
a \code{git_blob} object
}
\description{
Lookup a blob object from a repository.
}
\keyword{methods}

