% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{lookup}
\alias{lookup}
\alias{lookup,git_repository-method}
\title{Lookup}
\usage{
lookup(object, hex)

\S4method{lookup}{git_repository}(object, hex)
}
\arguments{
\item{object}{The repository \code{object}.}

\item{hex}{the identity of the object to lookup}
}
\value{
a \code{git_blob} or \code{git_commit} or \code{git_tag}
or \code{git_tree} object
}
\description{
Lookup one object in a repository.
}
\keyword{methods}

