% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,git_branch-method}
\title{Merge a branch into HEAD}
\usage{
merge(object, ...)

\S4method{merge}{git_branch}(object, commit_on_success = TRUE)
}
\arguments{
\item{object}{A \code{\linkS4class{git_branch}} object to merge
into HEAD.}

\item{...}{Additional arguments affecting the merge}

\item{commit_on_success}{If there are no conflicts written to the
index, the merge commit will be committed. Default is TRUE.}
}
\value{
A \code{\linkS4class{git_merge_result}} object.
}
\description{
Merge a branch into HEAD
}
\keyword{methods}

