% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{merge_base}
\alias{merge_base}
\alias{merge_base,git_repository,git_commit,git_commit-method}
\title{Find a merge base between two commits}
\usage{
merge_base(repo, one, two)

\S4method{merge_base}{git_repository,git_commit,git_commit}(repo, one, two)
}
\arguments{
\item{repo}{The repository}

\item{one}{One of the commits}

\item{two}{The other commit}
}
\value{
S4 class git_commit
}
\description{
Find a merge base between two commits
}
\keyword{methods}

