% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Optimise a vector for storage in to a git repository and add meta data}
\usage{
meta(x)
}
\arguments{
\item{x}{the vector}
}
\description{
Optimise a vector for storage in to a git repository and add meta data
}
