% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{note_create}
\alias{note_create}
\alias{note_create,git_blob-method}
\alias{note_create,git_commit-method}
\alias{note_create,git_tree-method}
\title{Add note for a object}
\usage{
note_create(object, message, ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)

\S4method{note_create}{git_blob}(object, message,
  ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)

\S4method{note_create}{git_commit}(object, message,
  ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)

\S4method{note_create}{git_tree}(object, message,
  ref = note_default_ref(object@repo),
  author = default_signature(object@repo),
  committer = default_signature(object@repo), force = FALSE)
}
\arguments{
\item{object}{The object to annotate}

\item{message}{Content of the note to add}

\item{ref}{Canonical name of the reference to use. Default is
\code{note_default_ref}.}

\item{author}{Signature of the notes note author}

\item{committer}{Signature of the notes note committer}

\item{force}{Overwrite existing note. Default is FALSE}
}
\value{
S4 class git_note
}
\description{
Add note for a object
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)

## Create a file, add and commit
writeLines("Hello world!", file.path(path, "test.txt"))
add(repo, "test.txt")
commit.1 <- commit(repo, "Commit message 1")

## Create another commit
writeLines(c("Hello world!",
             "HELLO WORLD!"),
           file.path(path, "test.txt"))
add(repo, "test.txt")
commit.2 <- commit(repo, "Commit message 2")

## Check that note_list is an empty list
stopifnot(identical(note_list(repo), list()))

## Create note in default namespace
note.1 <- note_create(commit.1, "Note-1")
note.2 <- note_create(commit.1, "Note-2", force = TRUE)

## Create note in named (review) namespace
note.3 <- note_create(commit.1, "Note-3", ref="refs/notes/review")
note.4 <- note_create(commit.2, "Note-4", ref="refs/notes/review")
note_remove(note.3)
note_remove(note.4)
note.5 <- note_create(commit.1, "Note-5", ref="review")
note.6 <- note_create(commit.2, "Note-6", ref="review")

## Create note on blob and tree
note.7 <- note_create(tree(commit.1), "Note-7")
note.8 <- note_create(tree(commit.1)["test.txt"], "Note-8")

## Cleanup
unlink(path, recursive=TRUE)
}
}
\keyword{methods}

