% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{note_list}
\alias{note_list}
\alias{note_list,git_repository,character-method}
\alias{note_list,git_repository,missing-method}
\title{List notes}
\usage{
note_list(repo, ref)

\S4method{note_list}{git_repository,missing}(repo)

\S4method{note_list}{git_repository,character}(repo, ref)
}
\arguments{
\item{repo}{The repository}

\item{ref}{Reference to read from. Default is
"refs/notes/commits".}
}
\value{
list with S4 class git_note objects
}
\description{
List all the notes within a specified namespace.
}
\examples{
\dontrun{
## Create and initialize a repository in a temporary directory
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)
config(repo, user.name="User", user.email="user@example.org")

## Create a file, add and commit
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit_1 <- commit(repo, "Commit message 1")

## Create another commit
writeLines(c("Hello world!",
             "HELLO WORLD!"),
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit_2 <- commit(repo, "Commit message 2")

## Create note in default namespace
note_create(commit_1, "Note-1")
note_create(commit_1, "Note-2", force = TRUE)

## Create note in named (review) namespace
note_create(commit_1, "Note-3", ref="refs/notes/review")
note_create(commit_2, "Note-4", ref="review")

## Create note on blob and tree
note_create(tree(commit_1), "Note-5")
note_create(tree(commit_1)["example.txt"], "Note-6")

## List notes in default namespace
note_list(repo)

## List notes in 'review' namespace
note_list(repo, "review")
}
}
\keyword{methods}

