% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{note_list}
\alias{note_list}
\alias{note_list,git_repository-method}
\title{List notes}
\usage{
note_list(repo, ref)

\S4method{note_list}{git_repository}(repo, ref)
}
\arguments{
\item{repo}{The repository}

\item{ref}{Reference to read from. Default is NULL, which uses
"refs/notes/commits".}
}
\value{
list with S4 class git_note objects
}
\description{
List all the notes within a specified namespace.
}
\keyword{methods}

