% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{pull}
\alias{pull}
\alias{pull,git_repository-method}
\title{Pull}
\usage{
pull(repo, merger = default_signature(repo))

\S4method{pull}{git_repository}(repo, merger = default_signature(repo))
}
\arguments{
\item{repo}{the repository}

\item{merger}{Who made the merge, if the merge is non-fast forward
merge that creates a merge commit.}
}
\value{
invisible(NULL)
}
\description{
Pull
}
\keyword{methods}

