% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punch_card.r
\docType{methods}
\name{punch_card}
\alias{punch_card}
\alias{punch_card,git_repository-method}
\title{Punch card}
\usage{
punch_card(repo, main = NULL, ...)

\S4method{punch_card}{git_repository}(repo, main = NULL, ...)
}
\arguments{
\item{repo}{The repository}

\item{main}{Default title for the plot is "Punch card on repo:"
and repository workdir basename. Supply a new title if you desire
one.}

\item{...}{Additional arguments affecting the plot}
}
\value{
invisible NULL
}
\description{
Punch card
}
\examples{
\dontrun{
## Initialize repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- clone("https://github.com/ropensci/git2r.git", path)

## Plot
punch_card(repo)
}
}
\keyword{methods}

