% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{push}
\alias{push}
\alias{push,git_branch-method}
\alias{push,git_repository-method}
\title{Push}
\usage{
push(object, ...)

\S4method{push}{git_branch}(object, credentials = NULL)

\S4method{push}{git_repository}(object, name = NULL, refspec = NULL,
  credentials = NULL)
}
\arguments{
\item{object}{S4 class \code{git_repository} or \code{git_branch}.}

\item{...}{Additional arguments affecting the push.}

\item{credentials}{The credentials for remote repository
access. Default is NULL.}

\item{name}{The remote's name. Default is NULL.}

\item{refspec}{The refspec to be pushed. Default is NULL.}
}
\value{
invisible(NULL)
}
\description{
Push
}
\seealso{
\code{\linkS4class{cred_plaintext}},
\code{\linkS4class{cred_ssh_key}}
}
\keyword{methods}

