% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim_git.R
\name{read_delim_file}
\alias{read_delim_file}
\title{Read a \code{data.frame} from a git repository}
\usage{
read_delim_file(file, repo)
}
\arguments{
\item{file}{the name of the file without file extension. Can include a relative
path. It is relative to the "project" when set in the \code{repo}. Otherwise
it is relative to the root of the \code{repo}.}

\item{repo}{a \code{git_repository} object, created with
\code{\link{repository}}}
}
\value{
The \code{data.frame}
}
\description{
Read a \code{data.frame} from a git repository
}
