% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R, R/merge.R, R/repository.R, R/tree.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{diff}
\alias{merge}
\alias{head}
\alias{as.data.frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{base}{\code{\link[base]{as.data.frame}}, \code{\link[base]{diff}}, \code{\link[base]{merge}}}

  \item{utils}{\code{\link[utils]{head}}}
}}

