% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{references}
\alias{references}
\alias{references,git_repository-method}
\title{Get all references that can be found in a repository.}
\usage{
references(repo)

\S4method{references}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository \code{object}.}
}
\value{
Character vector with references
}
\description{
Get all references that can be found in a repository.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## List all references in repository
references(repo)
}
}
\keyword{methods}

