% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reflog}
\alias{reflog}
\alias{reflog,git_repository-method}
\title{List and view reflog information}
\usage{
reflog(repo, refname = "HEAD", verbose = TRUE)

\S4method{reflog}{git_repository}(repo, refname = "HEAD", verbose = TRUE)
}
\arguments{
\item{repo}{S4 class git_repository \code{object}.}

\item{refname}{The name of the reference to list. 'HEAD' by
default.}

\item{verbose}{Display reflog. Default TRUE.}
}
\value{
invisible list of S4 git_reflog_entry objects
}
\description{
List and view reflog information
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

reflog(repo)
}
}
\keyword{methods}

