% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{reflog}
\alias{reflog}
\alias{reflog,git_repository-method}
\title{List and view reflog information}
\usage{
reflog(repo, refname = "HEAD", verbose = TRUE)

\S4method{reflog}{git_repository}(repo, refname = "HEAD", verbose = TRUE)
}
\arguments{
\item{repo}{S4 class git_repository \code{object}.}

\item{refname}{The name of the reference to list. 'HEAD' by
default.}

\item{verbose}{Display reflog. Default TRUE.}
}
\value{
invisible list of S4 git_reflog_entry objects
}
\description{
List and view reflog information
}
\examples{
\dontrun{
## Initialize a repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Config user
config(repo, user.name="Author", user.email="author@example.org")

## Write to a file and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Change file and commit
writeLines(c("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
             "eiusmod tempor incididunt ut labore et dolore magna aliqua."),
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "Second commit message")

## Change file again and commit
writeLines(c("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
             "eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad",
             "minim veniam, quis nostrud exercitation ullamco laboris nisi ut"),
           file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "Third commit message")

## View reflog
reflog(repo)
}
}
\keyword{methods}

