% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remote_add}
\alias{remote_add}
\alias{remote_add,git_repository-method}
\title{Add a remote to a repo}
\usage{
remote_add(object, ...)

\S4method{remote_add}{git_repository}(object, name, url)
}
\arguments{
\item{object}{The repository \code{object} to add the remote to}

\item{name}{Short name of the remote repository}

\item{url}{URL of the remote repository}
}
\value{
NULL, invisibly
}
\description{
Add a remote to a repo
}
\keyword{methods}

