% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remote_remove}
\alias{remote_remove}
\alias{remote_remove,git_repository,character-method}
\title{Remove a remote}
\usage{
remote_remove(repo, name)

\S4method{remote_remove}{git_repository,character}(repo, name)
}
\arguments{
\item{repo}{The repository to work on}

\item{name}{The name of the remote to remove}
}
\value{
NULL, invisibly
}
\description{
All remote-tracking branches and configuration settings for the
remote will be removed.
}
\keyword{methods}

