% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.r
\name{remote_rename}
\alias{remote_rename}
\title{Rename a remote}
\usage{
remote_rename(repo = NULL, oldname = NULL, newname = NULL)
}
\arguments{
\item{repo}{a \code{\linkS4class{git_repository}} object, or a
path to a repository, or \code{NULL}.  If the \code{repo}
argument is \code{NULL}, the repository is searched for with
\code{\link{discover_repository}} in the current working
directory.}

\item{oldname}{Old name of the remote}

\item{newname}{New name of the remote}
}
\value{
NULL, invisibly
}
\description{
Rename a remote
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user and commit a file
config(repo, user.name="Alice", user.email="alice@example.org")
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Add a remote
remote_add(repo, "playground", "https://example.org/git2r/playground")
remotes(repo)
remote_url(repo, "playground")

## Rename a remote
remote_rename(repo, "playground", "foobar")
remotes(repo)
remote_url(repo, "foobar")

## Set remote url
remote_set_url(repo, "foobar", "https://example.org/git2r/foobar")
remotes(repo)
remote_url(repo, "foobar")

## Remove a remote
remote_remove(repo, "foobar")
remotes(repo)
}
}
