% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{remote_url}
\alias{remote_url}
\alias{remote_url,git_repository-method}
\title{Get the remote url for remotes in a repo}
\usage{
remote_url(object, remote = remotes(object))

\S4method{remote_url}{git_repository}(object, remote = remotes(object))
}
\arguments{
\item{object}{The repository \code{object} to check remote_url}

\item{remote}{:TODO:DOCUMENTATION:}
}
\value{
Character vector with remote_url
}
\description{
Get the remote url for remotes in a repo
}
\keyword{methods}

