\docType{methods}
\name{remote_url-methods}
\alias{remote_url}
\alias{remote_url-methods}
\alias{remote_url,repository-method}
\title{Get the remote url for remotes in a repo}
\arguments{
  \item{object}{The repository \code{object} to check
  remote_url}
}
\value{
  Character vector with remote_url
}
\description{
  Get the remote url for remotes in a repo
}
\keyword{methods}

