\docType{methods}
\name{remotes-methods}
\alias{remotes}
\alias{remotes-methods}
\alias{remotes,repository-method}
\title{Get the configured remotes for a repo}
\arguments{
  \item{object}{The repository \code{object} to check
  remotes}
}
\value{
  Character vector with remotes
}
\description{
  Get the configured remotes for a repo
}
\keyword{methods}

