% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{remotes-methods}
\alias{remotes}
\alias{remotes-methods}
\title{Get the configured remotes for a repo}
\usage{
remotes(object)
}
\arguments{
\item{object}{The repository \code{object} to check remotes}
}
\value{
Character vector with remotes
}
\description{
Get the configured remotes for a repo
}
\keyword{methods}

