% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repository.r
\docType{methods}
\name{repository}
\alias{repository}
\alias{repository,character-method}
\title{Open a repository}
\usage{
repository(path, discover = FALSE)

\S4method{repository}{character}(path, discover = FALSE)
}
\arguments{
\item{path}{A path to an existing local git repository}

\item{discover}{Discover repository from path. Default is FALSE.}
}
\value{
A S4 \code{\linkS4class{git_repository}} object
}
\description{
Open a repository
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

# Configure a user
config(repo, user.name="Alice", user.email="alice@example.org")

## Create a file, add and commit
writeLines("Hello world!", file.path(path, "test-1.txt"))
add(repo, 'test-1.txt')
commit_1 <- commit(repo, "Commit message")

## Make one more commit
writeLines(c("Hello world!", "HELLO WORLD!"), file.path(path, "test-1.txt"))
add(repo, 'test-1.txt')
commit(repo, "Next commit message")

## Create one more file
writeLines("Hello world!", file.path(path, "test-2.txt"))

## Brief summary of repository
repo

## Summary of repository
summary(repo)

## Workdir of repository
workdir(repo)

## Check if repository is bare
is_bare(repo)

## Check if repository is empty
is_empty(repo)

## Check if repository is a shallow clone
is_shallow(repo)

## List all references in repository
references(repo)

## List all branches in repository
branches(repo)

## Get HEAD of repository
head(repo)

## Check if HEAD is head
is_head(head(repo))

## Check if HEAD is local
is_local(head(repo))

## List all tags in repository
tags(repo)
}
}
\keyword{methods}

