% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_commit-method}
\alias{show,git_commit-method}
\alias{show,git_commit-methods}
\title{Brief summary of commit}
\usage{
\S4method{show}{git_commit}(object)
}
\arguments{
\item{object}{The commit \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Displays the first seven characters of the hex, the date and the
summary of the commit message:
\code{[shortened hex] yyyy-mm-dd: summary}
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Brief summary of commits in repository
invisible(lapply(commits(repo), show))
}
}
\keyword{methods}

