% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,git_diff-method}
\alias{show,git_diff-method}
\alias{show,git_diff-methods}
\title{Show a diff}
\usage{
\S4method{show}{git_diff}(object)
}
\arguments{
\item{object}{The diff \code{object}.}
}
\description{
Show a diff
}
\keyword{methods}

