% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{show,git_repository-method}
\alias{show,git_repository-method}
\alias{show,git_repository-methods}
\title{Brief summary of repository}
\usage{
\S4method{show}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of repository
}
\keyword{methods}

