% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.r
\docType{methods}
\name{show,git_signature-method}
\alias{show,git_signature-method}
\alias{show,git_signature-methods}
\title{Brief summary of signature}
\usage{
\S4method{show}{git_signature}(object)
}
\arguments{
\item{object}{The repository \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of signature
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user
config(repo, user.name="Alice", user.email="alice@example.org")

## Brief summary of default signature
default_signature(repo)
}
}
\keyword{methods}
