% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,git_signature-method}
\alias{show,git_signature-method}
\alias{show,git_signature-methods}
\title{Brief summary of signature}
\usage{
\S4method{show}{git_signature}(object)
}
\arguments{
\item{object}{The repository \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of signature
}
\keyword{methods}

