% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{show,git_time-method}
\alias{show,git_time-method}
\alias{show,git_time-methods}
\title{Brief summary of branch}
\usage{
\S4method{show}{git_time}(object)
}
\arguments{
\item{object}{The time \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of branch
}
\keyword{methods}

