% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\docType{methods}
\name{show,git_time-method}
\alias{show,git_time-method}
\alias{show,git_time-methods}
\title{Brief summary of \code{git_time}}
\usage{
\S4method{show}{git_time}(object)
}
\arguments{
\item{object}{The time \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of \code{git_time}
}
\examples{
\dontrun{
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user
config(repo, user.name="Alice", user.email="alice@example.org")

## Brief summary of git_time from the default signature
default_signature(repo)@when
}
}
\keyword{methods}
