% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{stash}
\alias{stash}
\alias{stash,git_repository-method}
\title{Stash}
\usage{
stash(object, message = as.character(Sys.time()), index = FALSE,
  untracked = FALSE, ignored = FALSE, stasher = default_signature(object))

\S4method{stash}{git_repository}(object, message = as.character(Sys.time()),
  index = FALSE, untracked = FALSE, ignored = FALSE,
  stasher = default_signature(object))
}
\arguments{
\item{object}{The repository \code{object}.}

\item{message}{Optional description. Defaults to current time.}

\item{index}{All changes already added to the index are left
intact in the working directory. Default is FALSE}

\item{untracked}{All untracked files are also stashed and then
cleaned up from the working directory. Default is FALSE}

\item{ignored}{All ignored files are also stashed and then cleaned
up from the working directory. Default is FALSE}

\item{stasher}{Signature with stasher and time of stash}
}
\value{
invisible S4 class git_commit
}
\description{
Stash
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Create stash in repository
stash(repo, "Stash message")
}
}
\keyword{methods}

