% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{stash_drop}
\alias{stash_drop}
\alias{stash_drop,git_repository-method}
\alias{stash_drop,git_stash-method}
\title{Drop stash}
\usage{
stash_drop(object, ...)

\S4method{stash_drop}{git_repository}(object, index)

\S4method{stash_drop}{git_stash}(object)
}
\arguments{
\item{object}{The stash \code{object} to drop or a zero-based
integer to the stash to drop. The last stash has index 0.}
}
\value{
invisible NULL
}
\description{
Drop stash
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Assuming there are stashes in the repository.
## Drop a stash in repository.
stash_drop(stashes(repo)[[1]])

## Assuming there are stashes in the repository.
## Drop last stash in repository.
stash_drop(repo, 0)
}
}
\keyword{methods}

