% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{stash_list}
\alias{stash_list}
\alias{stash_list,git_repository-method}
\title{List stashes in repository}
\usage{
stash_list(repo)

\S4method{stash_list}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository.}
}
\value{
list of stashes in repository
}
\description{
List stashes in repository
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## List stashes in repository
stash_list(repo)
}
}
\keyword{methods}

