% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{status}
\alias{status}
\alias{status,git_repository-method}
\title{Status}
\usage{
status(repo, staged = TRUE, unstaged = TRUE, untracked = TRUE,
  ignored = FALSE)

\S4method{status}{git_repository}(repo, staged = TRUE, unstaged = TRUE,
  untracked = TRUE, ignored = FALSE)
}
\arguments{
\item{repo}{the \code{git_repository} to get status from.}

\item{staged}{include staged files. Default TRUE.}

\item{unstaged}{include unstaged files. Default TRUE.}

\item{untracked}{include untracked files. Default TRUE.}

\item{ignored}{include ignored files. Default FALSE.}
}
\value{
invisible(list) with repository status
}
\description{
Display state of the repository working directory and the staging
area.
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

status(repo)
}
}
\keyword{methods}

