% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,git_blob-method}
\alias{summary,git_blob-method}
\title{Summary of blob}
\usage{
\S4method{summary}{git_blob}(object, ...)
}
\arguments{
\item{object}{The blob \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of blob
}
\keyword{methods}

