% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,git_commit-method}
\alias{summary,git_commit-method}
\alias{summary,git_commit-methods}
\title{Summary of commit}
\usage{
\S4method{summary}{git_commit}(object, ...)
}
\arguments{
\item{object}{The commit \code{object}}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of commit
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Summary of commit in repository
summary(commits(repo)[[1]])
}
}
\keyword{methods}

