% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,git_stash-method}
\alias{summary,git_stash-method}
\alias{summary,git_stash-methods}
\title{Summary of a stash}
\usage{
\S4method{summary}{git_stash}(object, ...)
}
\arguments{
\item{object}{The stash \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of a stash
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Apply summary to each stash in the repository
invisible(lapply(stashes(repo), summary))
}
}
\keyword{methods}

