% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,git_tree-method}
\alias{summary,git_tree-method}
\alias{summary,git_tree-methods}
\title{Summary of tree}
\usage{
\S4method{summary}{git_tree}(object, ...)
}
\arguments{
\item{object}{The tree \code{object}}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of tree
}
\keyword{methods}

