% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tags}
\alias{tags}
\alias{tags,git_repository-method}
\title{Tags}
\usage{
tags(object)

\S4method{tags}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object}.}
}
\value{
list of tags in repository
}
\description{
Tags
}
\keyword{methods}

