% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,git_tree,ANY,missing-method}
\alias{[,git_tree,ANY,missing-method}
\title{Extract object from tree}
\usage{
\S4method{[}{git_tree,ANY,missing}(x, i)
}
\arguments{
\item{x}{The tree \code{object}}

\item{i}{The index of the tree object to extract}
}
\value{
Git object
}
\description{
Lookup a tree entry by its position in the tree
}
\keyword{methods}

