% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,git_tree,integer,missing-method}
\alias{[,git_tree,character,missing-method}
\alias{[,git_tree,integer,missing-method}
\alias{[,git_tree,logical,missing-method}
\alias{[,git_tree,numeric,missing-method}
\title{Extract object from tree}
\usage{
\S4method{[}{git_tree,integer,missing}(x, i)

\S4method{[}{git_tree,numeric,missing}(x, i)

\S4method{[}{git_tree,logical,missing}(x, i)

\S4method{[}{git_tree,character,missing}(x, i)
}
\arguments{
\item{x}{The tree \code{object}}

\item{i}{The index (integer or logical) of the tree object to
extract. If negative values, all elements except those indicated
are selected. A character vector to match against the names of
objects to extract.}
}
\value{
Git object
}
\description{
Lookup a tree entry by its position in the tree
}
\examples{
\dontrun{
## Open an existing repository
repo <- repository("path/to/git2r")

## Pick a tree in the repository
tree_object <- tree(commits(repo)[[1]])

## Summarize tree
summary(tree_object)

## Select item by name
tree_object[".Rbuildignore"]

## Select first item in tree
tree_object[1]

## Select first three items in tree
tree_object[1:3]

## Select all blobs in tree
tree_object[sapply(as(tree_object, 'list'), is_blob)]
}
}
\keyword{methods}

