% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tree}
\alias{tree}
\alias{tree,git_commit-method}
\alias{tree,git_stash-method}
\title{Tree}
\usage{
tree(object)

\S4method{tree}{git_commit}(object)

\S4method{tree}{git_stash}(object)
}
\arguments{
\item{object}{the \code{commit} or \code{stash} object}
}
\value{
A S4 class git_tree object
}
\description{
Get the tree pointed to by a commit or stash.
}
\keyword{methods}

