% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{workdir}
\alias{workdir}
\alias{workdir,git_repository-method}
\title{Workdir of repository}
\usage{
workdir(object)

\S4method{workdir}{git_repository}(object)
}
\arguments{
\item{object}{The repository \code{object} to check workdir}
}
\value{
Character vector with workdir. If the repository is bare,
\code{NULL} will be returned.
}
\description{
Workdir of repository
}
\keyword{methods}

