\docType{methods}
\name{workdir-methods}
\alias{workdir}
\alias{workdir-methods}
\alias{workdir,repository-method}
\title{Workdir of repository}
\arguments{
  \item{object}{The repository \code{object} to check
  workdir}
}
\value{
  Character vector with workdir
}
\description{
  Workdir of repository
}
\keyword{methods}

