% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{workdir}
\alias{workdir}
\alias{workdir,git_repository-method}
\title{Workdir of repository}
\usage{
workdir(repo)

\S4method{workdir}{git_repository}(repo)
}
\arguments{
\item{repo}{The repository \code{object}.}
}
\value{
Character vector with the path of the workdir. If the
repository is bare, \code{NULL} will be returned.
}
\description{
Workdir of repository
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)

## Get the path of the workdir for repository
workdir(repo)
}
}
\keyword{methods}

