% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{workdir-methods}
\alias{workdir}
\alias{workdir-methods}
\title{Workdir of repository}
\usage{
workdir(object)
}
\arguments{
\item{object}{The repository \code{object} to check workdir}
}
\value{
Character vector with workdir
}
\description{
Workdir of repository
}
\keyword{methods}

