% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.r
\name{workdir}
\alias{workdir}
\title{Workdir of repository}
\usage{
workdir(repo = NULL)
}
\arguments{
\item{repo}{repository \code{object}
\code{\linkS4class{git_repository}}, or a path to a
repository, or \code{NULL}.  If the \code{repo} argument is
\code{NULL}, the repository is searched for with
\code{\link{discover_repository}} in the current working
directory.}
}
\value{
Character vector with the path of the workdir. If the
repository is bare, \code{NULL} will be returned.
}
\description{
Workdir of repository
}
\examples{
\dontrun{
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)

## Get the path of the workdir for repository
workdir(repo)
}
}
