% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim_git.R
\name{write_delim_git}
\alias{write_delim_git}
\title{Write a \code{data.frame} to a git repository}
\usage{
write_delim_git(x, file, repo = ".", force = FALSE)
}
\arguments{
\item{x}{the \code{data.frame}}

\item{file}{the name of the file without file extension. Can include a relative
path. It is relative to the "project" when set in the \code{repo}. Otherwise
it is relative to the root of the \code{repo}.}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{force}{Add ignored files. Default is FALSE.}
}
\value{
The relative path to the file
}
\description{
This will create two files. The \code{".tsv"} file contains the raw data.
The \code{".yml"} contains the meta data on the columns in YAML format.
}
