% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim_git.r
\docType{methods}
\name{write_delim_git}
\alias{write_delim_git}
\alias{write_delim_git,ANY,ANY,ANY-method}
\alias{write_delim_git,git_connection-methods}
\alias{write_delim_git,ANY,ANY,git_connection-method}
\alias{write_delim_git,git_connection-methods}
\title{Write a dataframe as a tab delimited file to a git repository and stage it}
\usage{
write_delim_git(x, file, connection, ...)

\S4method{write_delim_git}{ANY,ANY,ANY}(x, file, connection, ...)

\S4method{write_delim_git}{ANY,ANY,git_connection}(x, file, connection, ...)
}
\arguments{
\item{x}{the data.frame}

\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{git_connection}
object}

\item{...}{parameters passed to \code{git_connection()} when
\code{connection} is a path. When \code{force} is available it is passed
to \code{add()}.}
}
\value{
the SHA1 of the file
}
\description{
The existing file will be overwritten.
}
\author{
Thierry Onkelinx, \email{thierry.onkelinx@inbo.be}
}
