% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim_git.R
\name{write_delim_git}
\alias{write_delim_git}
\title{Write a \code{data.frame} to a git repository}
\usage{
write_delim_git(x, file, repo)
}
\arguments{
\item{x}{the \code{data.frame}}

\item{file}{the name of the file with file extension. Can include a path
relative to the path of the \code{repo}}

\item{repo}{a \code{git_repository} object, created with
\code{\link{repository}}}
}
\value{
The relative path to the file
}
\description{
This will create two files. The \code{".tsv"} file contains the raw data.
The \code{".yml"} contains the meta data on the columns in YAML format.
}
