% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repository.R
\name{write_delim_git}
\alias{write_delim_git}
\title{Write a \code{data.frame} to a git repository}
\usage{
write_delim_git(x, file, repo = ".", sorting, override = FALSE,
  stage = FALSE, force = FALSE)
}
\arguments{
\item{x}{the \code{data.frame}}

\item{file}{the name of the file without file extension. Can include a relative
path. It is relative to the "project" when set in the \code{repo}. Otherwise
it is relative to the root of the \code{repo}.}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{sorting}{a vector of column names defining which columns to use for
sorting \code{x} and in what order to use them. Defaults to
\code{colnames(x)}}

\item{override}{Ignore existing meta data. This is required when new
variables are added or variables are deleted. Setting this to TRUE can
potentially lead to large diffs. Defaults to FALSE.}

\item{stage}{immediatly stage the changes}

\item{force}{Add ignored files. Default is FALSE.}
}
\value{
a named vector with the hashes of the files. The names contains the
files with their paths relative to the root of the git_repository.
}
\description{
This will create two files. The \code{".tsv"} file contains the raw data.
The \code{".yml"} contains the meta data on the columns in YAML format.
}
